% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{adjust_FM_covariate}
\alias{adjust_FM_covariate}
\title{Adjusting functions for covariates}
\usage{
adjust_FM_covariate(Y, X, pos = NULL, thresh_lowcount = 0)
}
\arguments{
\item{Y}{integer, corresponding to log2 of the signal length. WARNING the ordering change for different values of s.
@param X matrix of containning the covaraite}
}
\description{
Simple function to adjust the observed curves for covariate (e.g. age, population structure)
}
