% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{affected_reg}
\alias{affected_reg}
\title{Extract coordinates of the regions affected by the different CS}
\usage{
affected_reg(obj)
}
\arguments{
\item{obj}{at fitted obj object}
}
\description{
Extract coordinates of the regions affected by the different CS
}
\details{
return a matrix with 3 columns. Each lines corresponds to a regions
in which the estimated credible bands are "crossing zero"/i.e. the effects are likely not to be 0 in this region.
the second column corresponds to the start of the region  and the third to the end of the affected region
}
