% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{Bhat_recov}
\alias{Bhat_recov}
\title{Transform raw regression coefficient to wavelet regression coefficients}
\usage{
Bhat_recov(Bhat, W1)
}
\arguments{
\item{Bhat}{matrix of regression coefficient}

\item{W1}{matrix associated with a wavelet transform}
}
\value{
Wavelet regression coefficient using wavelet transform from W1
}
\description{
Transform raw regression coefficient to wavelet regression coefficients
}
