% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin_Y}
\alias{bin_Y}
\title{Binned reponsed data}
\usage{
bin_Y(Y, base_pair, n_bins = 1024)
}
\arguments{
\item{Y}{a matrix of observed individual function measure in more than 1024 positions}

\item{base_pair}{the corresponding position of each column}

\item{n_bins}{number of bins for the ouput}
}
\description{
Binned reponsed data
}
