% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{cal_Bhat_Shat_ss}
\alias{cal_Bhat_Shat_ss}
\title{@title Compute wavelet summary stat}
\usage{
cal_Bhat_Shat_ss(susiF_ss.obj, data, partial = TRUE, ...)
}
\description{
@title Compute wavelet summary stat
}
