% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{cal_cor_cs}
\alias{cal_cor_cs}
\title{Compute correlation between credible sets}
\usage{
cal_cor_cs(susiF.obj, X)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{X}{matrix of covariates}
}
\description{
Compute correlation between credible sets
}
