% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{cal_expected_residual}
\alias{cal_expected_residual}
\alias{cal_expected_residual.susiF_ss}
\title{Compute expected residuals for susiF_ss.obj}
\usage{
cal_expected_residual(susiF_ss.obj, data, ...)

\method{cal_expected_residual}{susiF_ss}(susiF_ss.obj, data)
}
\arguments{
\item{susiF_ss.obj}{susiF_ss.obj object}

\item{data}{an object of the class suff_stat define by function \code{\link{make_data_suff_stat}}}
}
\value{
an update suff_stat object
}
\description{
Compute expected residuals for susiF_ss.obj
}
