% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{cal_KL_l}
\alias{cal_KL_l}
\alias{cal_KL_l.susiF}
\title{Compute KL divergence effect l}
\usage{
cal_KL_l(susiF.obj, l, Y, X, ...)

\method{cal_KL_l}{susiF}(susiF.obj, l, Y, X, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{l}{effect to update}

\item{Y}{Matrix of outcomes}

\item{X}{Matrix of covariates}
}
\value{
susiF object
}
\description{
Compute KL divergence effect l
}
