% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{cal_KL_l.susiF}
\alias{cal_KL_l.susiF}
\title{Compute KL divergence for effect l (correct ELBO version)}
\usage{
\method{cal_KL_l}{susiF}(obj, l, X, D, C, indx_lst, ...)
}
\description{
Compute KL divergence for effect l (correct ELBO version)
}
\keyword{internal}
