% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{cal_KL_l_ss}
\alias{cal_KL_l_ss}
\title{@title Compute KL summary stat}
\usage{
cal_KL_l_ss(susiF_ss.obj, l, data, ...)
}
\arguments{
\item{susiF_ss.obj}{an object of class susiF_ss}

\item{l}{integer , effect of interest}

\item{data}{a data object see init_susiF_ss}
}
\description{
@title Compute KL summary stat
}
