% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{cal_L_mixsq_s_per_scale}
\alias{cal_L_mixsq_s_per_scale}
\title{Sub routine to compute likelihood matrix at scale s for mixsqp under mixture normal per scale prior}
\usage{
cal_L_mixsq_s_per_scale(G_prior, s, Bhat, Shat, indx_lst)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{s}{scale where the likelihood matrix should be computed}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{indx_list}{list generated by gen_wavelet_indx for the given level of resolution, used only with class \class{mixture_normal_per_scale}}
}
\value{
L see L argument mixsqp package mixsqp function
}
\description{
Sub routine to compute likelihood matrix at scale s for mixsqp under mixture normal per scale prior
}
