% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{cal_lfsr}
\alias{cal_lfsr}
\title{Compute conditional local false sign rate}
\usage{
cal_lfsr(clfsr_wc, alpha)
}
\arguments{
\item{clfsr_wc}{an object of the class "clfsr_wc" generated by \code{\link{cal_clfsr}}}

\item{alpha}{vector of length P containning}
}
\value{
J local false sign rate
}
\description{
Compute conditional local false sign rate
}
\keyword{internal}
