% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{cal_lik}
\alias{cal_lik}
\title{Compute likelihood for the weighted ash problem}
\usage{
cal_lik(lBF, zeta)
}
\arguments{
\item{lBF}{vector of log Bayes factors}

\item{zeta}{assignment probabilities}
}
\value{
Likelihood value
}
\description{
Add description here.
}
\keyword{internal}
