% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_EBmvFR_obj.R,
%   R/operation_on_susiF_obj.R
\name{cal_partial_resid.EBmvFR}
\alias{cal_partial_resid.EBmvFR}
\alias{cal_partial_resid}
\alias{cal_partial_resid.susiF}
\title{Compute partial residual for effect l}
\usage{
\method{cal_partial_resid}{EBmvFR}(obj, l, X, D, C, indx_lst, ...)

cal_partial_resid(susiF.obj, l, X, D, C, indx_lst, ...)

\method{cal_partial_resid}{susiF}(susiF.obj, l, X, D, C, indx_lst, ...)
}
\arguments{
\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{X}{matrix of covariates}

\item{D}{matrix of wavelet D coefficients from the original input data (Y)}

\item{C}{vector of wavelet scaling coefficient from the original input data (Y)}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{susiF.obj}{a susiF object defined by init_susiF_obj function}
}
\value{
a matrix of size N by size J of partial residuals
}
\description{
Compute partial residual for effect l
}
\keyword{internal}
