% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{cal_partial_resid.susiF}
\alias{cal_partial_resid.susiF}
\title{Compute partial residual for effect l}
\usage{
\method{cal_partial_resid}{susiF}(susiF.obj, l, X, D, C, L, indx_lst)
}
\arguments{
\item{susiF.obj}{a susisf object defined by init_susiF_obj function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{X}{matrix of covariate}

\item{D}{matrix of wavelet D coefficients from the original input data (Y)}

\item{C}{vector of wavelet scaling coefficient from the original input data (Y)}

\item{L}{total number of effect}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
a matrix of size N by size J of partial residuals
}
\description{
Compute partial residual for effect l
}
