% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cal_purity}
\alias{cal_purity}
\title{Compute purity of a list of CS}
\usage{
cal_purity(l_cs, X)
}
\arguments{
\item{l_cs}{the list of Credible sets}

\item{X}{the regression matrix}
}
\description{
Compute purity
}
\keyword{internal}
