% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{change_fit}
\alias{change_fit}
\title{Change postprocessing used in susiF object}
\usage{
change_fit(
  obj,
  Y,
  X,
  to = "TI",
  verbose = TRUE,
  max_scale = 10,
  filter_cs = FALSE,
  filter.number = 10,
  family = "DaubLeAsymm"
)
}
\arguments{
\item{obj}{a fitted susiF  object}

\item{Y}{functional phenotype used in the original fit}

\item{X}{matrix of  covariates  used in the original fit}

\item{to}{name of the type of postprocessing you wish to change the susiF  object  to be fitted with
current option "TI" or "hMM"}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
and a summary of the optimization settings are printed on the
console.}

\item{max_scale}{numeric, define the maximum of wavelet coefficients used in the analysis (2^max_scale).
Set 10 true by default.}

\item{filter_cs}{logical, if TRUE filters the credible set (removing low-purity)
cs and cs with estimated prior equal to 0). Set as TRUE by default.}

\item{filter.number}{see documentation of wd from wavethresh package}

\item{family}{see documentation of wd from wavethresh package}
}
\description{
Change postprocessing used in susiF object
}
