% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{check_cs}
\alias{check_cs}
\title{Check purity credible sets}
\usage{
check_cs(susiF.obj, min.purity = 0.5)
}
\arguments{
\item{susiF.obj}{a susif object defined by \code{\link{init_susiF_obj}} function}
}
\value{
a susif.obj without "dummy" credible sets
}
\description{
Check purity credible sets
}
