% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colScale}
\alias{colScale}
\title{Scaling function from r-blogger}
\usage{
colScale(
  x,
  center = TRUE,
  scale = TRUE,
  add_attr = TRUE,
  rows = NULL,
  cols = NULL
)
}
\arguments{
\item{x}{a matrix}

\item{center}{logical if true center column}

\item{scale}{logical if true scale column}

\item{add_attr}{logical  https://www.r-bloggers.com/2016/02/a-faster-scale-function/}

\item{rows}{logical  https://www.r-bloggers.com/2016/02/a-faster-scale-function/}

\item{cols}{logical  https://www.r-bloggers.com/2016/02/a-faster-scale-function/}
}
\description{
from https://www.r-bloggers.com/2016/02/a-faster-scale-function/
}
