% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{DWT2}
\alias{DWT2}
\title{wavelet transform on matrix of functions}
\usage{
DWT2(data, filter.number = 10, family = "DaubLeAsymm")
}
\arguments{
\item{data}{matrix of size NxJ where J is a power of two}

\item{filter.number}{This selects the smoothness of wavelet that you want to use in the decomposition. By default this is 10, the Daubechies least-asymmetric orthonormal compactly supported wavelet with 10 vanishing moments. Description from Guy Nason (wavethresh package)}

\item{family}{specifies the family of wavelets that you want to use. Two popular options are "DaubExPhase" and "DaubLeAsymm" but see the help for filter.select for more possibilities.. Description from Guy Nason (wavethresh package)}
}
\value{
A list with the following components

\item{C}{ vector length n containing the wavelet C coefficient}
\item{D}{ matrix of size nx 2^(J -1) where each row contains the wavelet D coefficients, ordered in the same way as in the wavethresh package}
\item{family}{ used for the wavelet transform}
\item{filter.number}{ }
}
\description{
function adapted from grove R package from Ma and Soriano. Perform wavelet transform of each row of a matrix
}
