% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_poisson_mean_routines.R
\name{ebpm_normal}
\alias{ebpm_normal}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
ebpm_normal(
  x,
  s = NULL,
  g_init = NULL,
  fix_g = FALSE,
  q_init = NULL,
  maxiter = 20,
  tol = 1e-05,
  vga_tol = 1e-05,
  conv_type = "sigma2abs",
  return_sigma2_trace = FALSE
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{g_init}{a list of mean, and var. Can be NULL for both parameters.}

\item{fix_g}{Whether fix g at g_init. If only fix either mean, or var, fix_g can be a length 2 boolean vector.}

\item{q_init}{a list of init value of m_init(posterior mean) and v_init(posterior var).}

\item{maxiter}{max number of iterations}

\item{tol}{tolerance for stopping the updates}

\item{conv_type}{convergence criteria, default to be elbo}

\item{return_sigma2_trace}{whether return the trace of sigma2 estiamtes}
}
\value{
a list of
 \item{posterior:}{mean_log/var_log is the posterior mean/var of mu, mean is the posterior of exp(mu)}
 \item{fitted_g:}{estimated prior}
 \item{obj_value:}{objective function values}
}
\description{
Gaussian prior, Gaussian posterior in Poisson mean problem.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(\beta,\sigma^2).}
}
\examples{
\dontrun{
n = 1000
mu = rnorm(n)
x = rpois(n,exp(mu))
ebpm_normal(x)
}
}
