% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ebps.R
\name{ebps_init_control_default}
\alias{ebps_init_control_default}
\title{Default parameters for ebps initialization}
\usage{
ebps_init_control_default()
}
\arguments{
\item{m_init_method}{'vga' or 'smash_poi'}
}
\description{
Default parameters for ebps initialization
}
