% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ebps.R
\name{ebps_smooth_control_default}
\alias{ebps_smooth_control_default}
\title{Default parameters for ebps smoothing function}
\usage{
ebps_smooth_control_default()
}
\arguments{
\item{ndwt_method}{'smash' or 'ti.thresh'}
}
\description{
Default parameters for ebps smoothing function
}
