% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{Eloglik}
\alias{Eloglik}
\alias{Eloglik.susiF}
\title{Expected log likelihood for a   susiF   object}
\usage{
Eloglik(obj, ...)

\method{Eloglik}{susiF}(obj, Y, X, ...)
}
\arguments{
\item{obj}{a susiF object defined by  init_susiF_obj  function}

\item{Y}{Matrix of outcomes}

\item{X}{Matrix of covariates}

\item{data}{list of data for susif_ss}
}
\value{
Expected log likelihood
}
\description{
Expected log likelihood for a   susiF   object
}
\keyword{internal}
