% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{EM_pi}
\alias{EM_pi}
\title{EM algorithm to find Empirical Bayes prior}
\usage{
EM_pi(G_prior, Bhat, Shat, indx_lst, max_step = 100, espsilon = 1e-04)
}
\arguments{
\item{G_prior}{mixture normal prior  or mixture  normal per scale}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{max_step}{numeric, maximum number of EM iteration}

\item{espsilon}{numeric, tolerance EM algorithm}

\item{indx_list}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class mixture_normal_per_scale}
}
\value{
\describe{
\item tpi_k fitted mixture proportion
\item lBF log Bayes Factor
}
}
\description{
EM algorithm to find Empirical Bayes prior
}
