% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{estimate_residual_variance}
\alias{estimate_residual_variance}
\title{Update residual variance}
\usage{
estimate_residual_variance(Y, X, susiF.obj)
}
\arguments{
\item{Y}{wavelet transformed  functional phenotype, matrix of size N by size J.}

\item{X}{matrix of size N by p}

\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}
}
\value{
estimated residual variance
}
\description{
Update residual variance
}
