% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{estimate_residual_variance_ss}
\alias{estimate_residual_variance_ss}
\title{Estimate residual variance for summary stat}
\usage{
estimate_residual_variance_ss(susiF_ss.obj, data, ...)
}
\arguments{
\item{susiF_ss.obj}{an object of class susiF_ss}

\item{data}{a data object see init_susiF_ss}
}
\description{
Estimate residual variance for summary stat
}
