% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{expand_susiF_obj}
\alias{expand_susiF_obj}
\title{Expand susiF.obj by adding L_extra effect}
\usage{
expand_susiF_obj(susiF.obj, L_extra)
}
\arguments{
\item{susiF.obj}{a susiF.obj}

\item{L_extra}{numeric a number of effect to add}
}
\value{
a susiF.obj a L_extra effect. Note the the number of effect of the susiF.obj cannot exceed the number the user upper bound
}
\description{
Expand susiF.obj by adding L_extra effect
}
