% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{fit_ash_level}
\alias{fit_ash_level}
\title{Fit ash of coefficient from scale s}
\usage{
fit_ash_level(Bhat, Shat, s, indx_lst, lowc_wc, ...)
}
\arguments{
\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{s}{scale of interest}

\item{indx_lst}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class mixture_normal_per_scale}

\item{lowc_wc}{wavelet coefficient with low count to be discarded}
}
\value{
an ash object
}
\description{
Add description here.
}
