% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{fit_lm}
\alias{fit_lm}
\title{Regress column l of Y on column j of X}
\usage{
fit_lm(l, j, Y, X, v1)
}
\arguments{
\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelet which assume that J is of the form J^2. If J not a power of 2, susif internally remaps the data into grid of length 2^J}

\item{X}{matrix of size n by p in}

\item{v1}{vector of 1 of length nn}
}
\value{
vector of 2 containing the regression coefficient and standard
}
\description{

}
