% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{fsusie_log_plot}
\alias{fsusie_log_plot}
\title{fSuSiE Plots using Gviz}
\usage{
fsusie_log_plot(
  obj,
  chr,
  pos0,
  pos1,
  X,
  Y,
  snp_info,
  cs = 1,
  log1p_count = FALSE,
  effect_log = TRUE,
  thresh_lfsr = NULL,
  data_splice = NULL,
  type_data = "p"
)
}
\arguments{
\item{obj}{is the fsusie object}

\item{chr}{the chromosome number}

\item{pos0}{the start of the RNAseq count (genomic position of the first column of
the Y matrix used to fit the fsusie object) description}

\item{pos1}{the end of the RNAseq count (genomic position of the last column of
the Y matrix used to fit the fsusie object) description}

\item{X}{the X matrix used to fit fsusie}

\item{Y}{the Y matrix used to fit fsusie}

\item{snp_info}{'optional) a matrix containing the information of the genotype matrix see vignette on RNaseq}

\item{cs}{the cs number to be plotted}

\item{log1p_count}{logical (default to FALSE) show the observe count conditional on the leads SNP in the
using log1p_count}

\item{effect_log}{logical (set to TRUE) , the plot assume that you fitted the fsusie object
on log +1 count and so if you set this parameter to FALSE the displayed effect will be the expected
difference in count instead of the fitted curve}

\item{thresh_lfsr}{if the susiF object is fitted using HMM postprocessing you can use this argument
to set to 0 the estimated effect that have an local false sign rate higher than a given threshold (e.g., 0.05) description}

\item{data_splice}{position of splicing sites
see GViz documentation}

\item{type_data}{set to "p" change the type of plot for the observed count conditional of the lead SNP}
}
\description{
fSuSiE Plots using Gviz
}
