% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_susiF}
\alias{plot_susiF}
\alias{plot.susiF}
\alias{plot_susiF_pip}
\alias{plot_susiF_effect}
\title{fSuSiE Plots}
\usage{
plot_susiF(
  obj,
  title = "",
  effect = "all",
  cred.band = TRUE,
  lfsr.curve = TRUE,
  linewidth = 0.5,
  point_size = 2,
  pos_SNP,
  point_shape,
  pip_only = FALSE
)

\method{plot}{susiF}(x, ...)

plot_susiF_pip(
  obj,
  title = "",
  point_size = 2,
  pos_SNP,
  point_shape,
  font_size = 10
)

plot_susiF_effect(
  obj,
  effect = "all",
  title = "",
  cred.band = TRUE,
  lfsr.curve = TRUE,
  linewidth = 0.5,
  font_size = 10
)
}
\arguments{
\item{obj}{Output of the susiF function.}

\item{title}{The title of the plot.}

\item{effect}{The indices of the effects to be plotted, or use
\code{effect = "all"} to plot all effects.}

\item{cred.band}{logical. If \code{TRUE}, plot credible bands if
the fSuSiE model was fitted with wavelet regression.}

\item{lfsr.curve}{Logical. If \code{TRUE}, plot estimated lfsr of the
effect at each base pair if obj fitted with HMM regression. This
has no effect unless the \code{\link{susiF}} option
\code{post_processing = "HMM"} was used.}

\item{linewidth}{Numeric. Width of the plotted lines.}

\item{point_size}{numeric, size of the point}

\item{pos_SNP}{vector, containing the base pair of the SNPs}

\item{point_shape}{vector, containing the shape of dots}

\item{pip_only}{logical, if TRUE only ouput the PIP plot}

\item{font_size}{Passed as the \dQuote{ont_size} argument to
\code{\link[cowplot]{theme_cowplot}}.}
}
\description{
Various visualizations of fSuSiE results.
}
