% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{get_alpha}
\alias{get_alpha}
\alias{get_alpha.susiF}
\alias{get_alpha.susiF_ss}
\title{Update alpha  susiF mixture proportion of effect l}
\usage{
get_alpha(susiF.obj, l, ...)

\method{get_alpha}{susiF}(susiF.obj, l, ...)

\method{get_alpha}{susiF_ss}(susiF.obj, l, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}
}
\value{
susiF object
}
\description{
Update alpha  susiF mixture proportion of effect l
}
