% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{get_ER2}
\alias{get_ER2}
\alias{get_ER2.susiF}
\title{Compute Epected sum of square}
\usage{
get_ER2(susiF.obj, Y, X, ...)

\method{get_ER2}{susiF}(susiF.obj, Y, X, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by init_susiF_obj function}

\item{Y}{wavelet transformed  functional phenotype, matrix of size N by size J.}

\item{X}{matrix of size N by p}
}
\value{
estimated residual variance
}
\description{
Compute Epected sum of square
}
\keyword{internal}
