% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{get_fitted_effect}
\alias{get_fitted_effect}
\alias{get_fitted_effect.susiF}
\title{Access fitted effect l}
\usage{
get_fitted_effect(obj, l, ...)

\method{get_fitted_effect}{susiF}(obj, l, ...)
}
\arguments{
\item{obj}{a fitted objected}

\item{l}{effect of interest}

\item{\dots}{Other arguments.}
}
\description{
Access fitted effect l
}
