% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{get_G_prior}
\alias{get_G_prior}
\alias{get_G_prior.susiF}
\alias{get_G_prior.susiF_ss}
\title{Access susiF internal prior}
\usage{
get_G_prior(susiF.obj, ...)

\method{get_G_prior}{susiF}(susiF.obj, ...)

\method{get_G_prior}{susiF_ss}(susiF_ss.obj, ...)
}
\arguments{
\item{susiF.obj}{a susisf object defined by \code{\link{init_susiF_obj}} function}
}
\value{
G_prior object
}
\description{
Access susiF internal prior
}
