% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{get_lBF}
\alias{get_lBF}
\title{Access susiF log Bayes factors of effect l}
\usage{
get_lBF(susiF.obj, l, ...)
}
\arguments{
\item{susiF.obj}{a susisf object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}
}
\value{
a vector of log Bayes Factors
}
\description{
Access susiF log Bayes factors of effect l
}
