% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{get_objective_ss}
\alias{get_objective_ss}
\title{@title Compute objective summary stat}
\usage{
get_objective_ss(susiF_ss.obj, data, ...)
}
\arguments{
\item{susiF_ss.obj}{an object of class susiF_ss}

\item{data}{a data object see init_susiF_ss}
}
\description{
@title Compute objective summary stat
}
