% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{get_pi}
\alias{get_pi}
\alias{get_pi.susiF}
\title{Access susiF mixture proportion of effect l}
\usage{
get_pi(susiF.obj, l, ...)

\method{get_pi}{susiF}(susiF.obj, l, ...)
}
\arguments{
\item{susiF.obj}{a susisf object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}
}
\value{
susiF object
}
\description{
Access susiF mixture proportion of effect l
}
