% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{get_pi0.mixture_normal}
\alias{get_pi0.mixture_normal}
\title{Get mixture proportion for mixture normal prior per scale}
\usage{
\method{get_pi0}{mixture_normal}(G_prior)
}
\arguments{
\item{G_prior}{mixture normal prior}
}
\value{
list of vector of mixture proportion
}
\description{
Add description here.
}
