% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{get_pi0.mixture_normal_per_scale}
\alias{get_pi0.mixture_normal_per_scale}
\title{Get mixture proportion for of the null component in mixture normal prior}
\usage{
\method{get_pi0}{mixture_normal_per_scale}(G_prior)
}
\arguments{
\item{G_prior}{mixture normal prior}
}
\value{
A number between 0 and 1
}
\description{
Get mixture proportion for of the null component in mixture normal prior
}
