% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_pi0.pi_mixture_normal}
\alias{get_pi0.pi_mixture_normal}
\title{Get mixture proportion for mixture normal prior}
\usage{
\method{get_pi0}{pi_mixture_normal}(tpi)
}
\arguments{
\item{tpi}{object of class pi_mixture_normal}
}
\value{
numeric between 0 an 1
}
\description{
Get mixture proportion for mixture normal prior
}
