% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{get_pi0.pi_mixture_normal_per_scale}
\alias{get_pi0.pi_mixture_normal_per_scale}
\title{Get mixture proportion for mixture normal prior}
\usage{
\method{get_pi0}{pi_mixture_normal_per_scale}(tpi)
}
\arguments{
\item{tpi}{object of class pi_mixture_normal}
}
\value{
A number between 0 and 1
}
\description{
Get mixture proportion for mixture normal prior
}
