% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_pi_G_prior.mixture_normal}
\alias{get_pi_G_prior.mixture_normal}
\title{Get mixture proportion for mixture normal prior}
\usage{
\method{get_pi_G_prior}{mixture_normal}(G_prior)
}
\arguments{
\item{G_prior}{mixture normal prior}
}
\value{
vector of mixture proportion
}
\description{
Get mixture proportion for mixture normal prior
}
