% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_pi_G_prior.mixture_normal_per_scale}
\alias{get_pi_G_prior.mixture_normal_per_scale}
\title{Get mixture proportion for mixture normal prior per scale}
\usage{
\method{get_pi_G_prior}{mixture_normal_per_scale}(G_prior)
}
\arguments{
\item{G_prior}{mixture normal prior}
}
\value{
list of vector of mixture proportion
}
\description{
Get mixture proportion for mixture normal prior per scale
}
