% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{get_pi_G_prior}
\alias{get_pi_G_prior}
\alias{get_pi_G_prior.mixture_normal}
\alias{get_pi_G_prior.mixture_normal_per_scale}
\title{Get mixture proportion for mixture prior}
\usage{
get_pi_G_prior(G_prior, ...)

\method{get_pi_G_prior}{mixture_normal}(G_prior, ...)

\method{get_pi_G_prior}{mixture_normal_per_scale}(G_prior, ...)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{\dots}{Other arguments.}
}
\value{
vector of mixture proportion
}
\description{
Get mixture proportion for mixture   prior
}
\keyword{internal}
