% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{get_sd_G_prior.mixture_normal}
\alias{get_sd_G_prior.mixture_normal}
\title{Get mixture standard deviations for mixture normal prior}
\usage{
\method{get_sd_G_prior}{mixture_normal}(G_prior)
}
\arguments{
\item{G_prior}{mixture normal prior}
}
\value{
vector of standard deviations
}
\description{
Add description here.
}
