% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get_sd_G_prior.mixture_normal_per_scale}
\alias{get_sd_G_prior.mixture_normal_per_scale}
\title{Get mixture standard deviations mixture normal prior per scale}
\usage{
\method{get_sd_G_prior}{mixture_normal_per_scale}(G_prior)
}
\arguments{
\item{G_prior}{mixture normal prior}
}
\value{
list of vectors of standard deviations
}
\description{
Get mixture standard deviations mixture normal prior per scale
}
