% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{get_sd_G_prior}
\alias{get_sd_G_prior}
\alias{get_sd_G_prior.mixture_normal}
\alias{get_sd_G_prior.mixture_normal_per_scale}
\title{Get mixture standard deviations for mixture normal prior}
\usage{
get_sd_G_prior(G_prior, ...)

\method{get_sd_G_prior}{mixture_normal}(G_prior, ...)

\method{get_sd_G_prior}{mixture_normal_per_scale}(G_prior, ...)
}
\arguments{
\item{G_prior}{mixture normal prior}
}
\value{
vector of standard deviations
}
\description{
Add description here.
}
