% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{greedy_backfit}
\alias{greedy_backfit}
\alias{greedy_backfit.susiF}
\title{Update  susiF via greedy search or backfit}
\usage{
greedy_backfit(obj, verbose, cov_lev, X, min.purity, ...)

\method{greedy_backfit}{susiF}(obj, verbose, cov_lev, X, min.purity, ...)
}
\arguments{
\item{obj}{a susiF object defined by init_susiF_obj function}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,}

\item{cov_lev}{the desired level of converage}

\item{X}{matrix of size n by p contains the covariates}

\item{min.purity}{minimum purity for estimated credible sets}
}
\value{
susiF object
}
\description{
Update  susiF via greedy search or backfit
}
\keyword{internal}
