% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HFT}
\alias{HFT}
\title{Perform Haar-Fisz tranform on  count matrix}
\usage{
HFT(count.data)
}
\arguments{
\item{count.data}{a N by T matrix of count data where each row is an observed Poisson processt}
}
\value{
a matrix of size N by T  of the Haar-Fisz transformed data
}
\description{
Perform Haar-Fisz tranform on  count matrix
}
