% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{HMM_regression}
\alias{HMM_regression}
\alias{HMM_regression.susiF}
\title{Compute refined estimate using HMM regression}
\usage{
HMM_regression(obj, Y, X, verbose = TRUE, fit_indval = TRUE, ...)

\method{HMM_regression}{susiF}(obj, Y, X, verbose = FALSE, fit_indval = TRUE, ...)
}
\arguments{
\item{obj}{a susiF object}

\item{Y}{matrix of responses}

\item{X}{matrix containing the covariates}

\item{verbose}{logical}

\item{fit_indval}{logical if set to true compute fitted value (default value TRUE)}

\item{\dots}{Other arguments.}
}
\description{
Compute refined estimate using HMM regression
}
