% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_EBmvFR_obj.R
\name{init_EBmvFR_obj}
\alias{init_EBmvFR_obj}
\title{Create an EBmvFR object}
\usage{
init_EBmvFR_obj(G_prior, Y, X, ..., lowc_wc = NULL)
}
\arguments{
\item{G_prior}{a prior generated vi the init_prior function}

\item{Y}{matrix of wavelet coefficients}

\item{X}{matrix of covariates}

\item{\dots}{Other arguments.}

\item{lowc_wc}{Wavelet columns excluded from fitting.}
}
\description{
Create an EBmvFR object
}
\details{
Create an EBmvFR object
}
