% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{init_prior}
\alias{init_prior}
\title{Initialize the prior}
\usage{
init_prior(...)
}
\value{
an object of the class "normal", "mixture_normal" or "mixture_normal_per_scale"
}
\description{
generate list of object corresponding to the parameters of the prior set for analysis
}
