% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{init_prior.suff_stat}
\alias{init_prior.suff_stat}
\title{Initialize the prior using sufficient satistics}
\usage{
\method{init_prior}{suff_stat}(data, prior, indx_lst)
}
\arguments{
\item{data}{an object of the class suff_stat define by function \code{\link{make_data_suff_stat}}}

\item{prior}{Three choice are available "normal", "mixture_normal", "mixture_normal_per_scale"}

\item{indx_lst}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class  mixture_normal_per_scale}
}
\value{
an object of the class "normal", "mixture_normal" or "mixture_normal_per_scale"
}
\description{
generate list of object corresponding to the parameters of the prior set for analysis
}
