% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{init_susiF_obj}
\alias{init_susiF_obj}
\title{Initialize a susiF object using regression coefficients}
\usage{
init_susiF_obj(L_max, G_prior, Y, X, L_start, greedy, backfit, ...)
}
\arguments{
\item{L_max}{upper bound on the number of non zero coefficients An L-vector containing the indices of the}

\item{G_prior}{prior object defined by init_prior function}

\item{Y}{Matrix of outcomes}

\item{X}{matrix of covariatess}

\item{L_start}{number of effect to start with}

\item{greedy}{logical, if TRUE allow greedy search}

\item{backfit}{logical, if TRUE allow backfitting}
}
\description{
Initialize a susiF object using regression coefficients
}
\details{
Initialize a susiF object using regression coefficients
}
