% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{interpol_mat}
\alias{interpol_mat}
\title{Lifting scheme for non decimated and unevenly spaced data on matrix}
\usage{
interpol_mat(Y, pos)
}
\arguments{
\item{Y}{matrix of curves of size nXT, where the curves are stored by row}

\item{pos}{vector of positive values, each the vector component corresponds to the measurement time. (the values of pos should increase along with the index)}
}
\description{
Interpolation procedure from Kovac and Silvermann 2000 for a matrix of functions
}
\details{
We suppose that we observe n functions/curves at T different time points. We remap these functions into a grid of length equal to a power of two using the lifting scheme of Kovac and Silvermann 2000.
}
