% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{interpolKS}
\alias{interpolKS}
\title{Interpolation procedure from Kovac and Silverman 2000}
\usage{
interpolKS(y, bp)
}
\arguments{
\item{y}{vector of observed curves}

\item{bp}{scaled observation time points on 0,1}
}
\description{
Interpolation procedure from Kovac and Silveramnn 2000
}
