% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{L_mixsq.mixture_normal_per_scale}
\alias{L_mixsq.mixture_normal_per_scale}
\title{Compute likelihood matrix for mixsqp under mixture  normal per scale prior}
\usage{
\method{L_mixsq}{mixture_normal_per_scale}(G_prior, Bhat, Shat, indx_lst)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{indx_list}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class  mixture_normal_per_scale}
}
\value{
List of L matrix (see L argument mixsqp package mixsqp function)
}
\description{
Add description here.
}
