% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{log_BF.mixture_normal_per_scale}
\alias{log_BF.mixture_normal_per_scale}
\title{Compute log Bayes Factor under mixture normal per scale prior}
\usage{
\method{log_BF}{mixture_normal_per_scale}(G_prior, Bhat, Shat, indx_lst, ...)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{indx_list}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class mixture_normal_per_scale}
}
\value{
p log Bayes factor
}
\description{
Add description here.
}
