% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{loglik_SFR_post_ss}
\alias{loglik_SFR_post_ss}
\title{Compute posterior expected log likelihood for
 single function regression of effect l for summary stat model}
\usage{
loglik_SFR_post_ss(susiF_ss.obj, l, data, ...)
}
\arguments{
\item{susiF_ss.obj}{an object of class susiF_ss}

\item{l}{effect to update}

\item{data}{a summary stat data format object}
}
\description{
Compute posterior expected log likelihood for
 single function regression of effect l for summary stat model
}
