% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{m_step.lik_mixture_normal}
\alias{m_step.lik_mixture_normal}
\title{Compute M step in the wieghted ash problem for normal mixture prior}
\usage{
\method{m_step}{lik_mixture_normal}(L, zeta, indx_lst)
}
\arguments{
\item{L}{output of L_mixsqp function}

\item{zeta}{assignment probabilities for each covariate}

\item{indx_list}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class  mixture_normal_per_scale}
}
\value{
a vector of proportion (class pi_mixture_normal)
}
\description{
Compute M step in the wieghted ash problem for normal mixture prior
}
