% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{m_step.lik_mixture_normal_per_scale}
\alias{m_step.lik_mixture_normal_per_scale}
\title{Compute M step in the weighted ash problem for normal mixture prior}
\usage{
\method{m_step}{lik_mixture_normal_per_scale}(L, zeta, indx_lst)
}
\arguments{
\item{L}{output of L_mixsqp function}

\item{zeta}{assignment probabilities for each covariate}
}
\value{
a list if vector of proportion (class pi_mixture_normal_per_scale)
}
\description{
Add description here.
}
