% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{make_data_suff_stat}
\alias{make_data_suff_stat}
\title{Define sufficient stat data format}
\usage{
make_data_suff_stat()
}
\arguments{
\item{Bhat}{A p by t matrix of estimated effects.}

\item{Shat}{A p by t matrix of standard errors.}

\item{R}{A p by p correlation matrix. It should be estimated from
  the same samples used to compute \code{bhat} and \code{shat}. Using
  an out-of-sample matrix may produce unreliable results.

 @param n The sample size.}

\item{var_y}{a T vector of The sample variance of Y at each time point , defined as \eqn{y'y/(n-1)}.
When the sample variance cannot be provided, the coefficients
(returned from \code{coef}) are computed on the "standardized" X, y
scale.}

\item{XtX}{A p by p matrix \eqn{X'X} in which the columns of X
are centered to have mean zero.}

\item{Xty}{A p by T matrix \eqn{X'y} in which y and the columns of X are
centered to have mean zero.}

\item{yty}{A T by T scalar \eqn{y'y} in which y is centered to have mean
zero.}
}
\description{
Define sufficient stat data format
}
