% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{merge_effect}
\alias{merge_effect}
\alias{merge_effect.susiF}
\title{Merging effect function}
\usage{
merge_effect(obj, verbose, ...)

\method{merge_effect}{susiF}(obj, verbose = FALSE, ...)
}
\arguments{
\item{obj}{a susiF object defined by init_susiF_obj function}

\item{verbose}{logical}

\item{\dots}{Other arguments.}
}
\value{
a susiF object
}
\description{
Merging effect function
}
\keyword{internal}
