% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{name_cs}
\alias{name_cs}
\alias{name_cs.susiF}
\title{Updates CS names for output}
\usage{
name_cs(susiF.obj, X, ...)

\method{name_cs}{susiF}(susiF.obj, X, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{X}{matrix of size N by p}
}
\description{
Updates CS names for output
}
