% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{out_prep.susiF}
\alias{out_prep.susiF}
\title{Preparing output of main susiF function}
\usage{
\method{out_prep}{susiF}(susiF.obj, Y, X, indx_lst)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelets that assume that J is of the form J^2. If J is not a power of 2, susiF internally remaps the data into a grid of length 2^J}

\item{X}{matrix of size n by p
#'@param indx_lst list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
susiF object
}
\description{
Preparing output of main susiF function
}
