% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{out_prep_ss}
\alias{out_prep_ss}
\title{Preparation output susiF_ss.obj type}
\usage{
out_prep_ss(susiF_ss.obj, ...)
}
\arguments{
\item{susiF_ss.obj}{an object of class susiF_ss}
}
\description{
Preparation output susiF_ss.obj type
}
