% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_effect}
\alias{plot_effect}
\title{Plot specific effect from susiF object}
\usage{
plot_effect(
  obj,
  effect = 1,
  title = "",
  cred.band = TRUE,
  lfsr.curve = TRUE,
  size_line = 2,
  ...
)
}
\arguments{
\item{obj}{output of the susiF function}

\item{effect}{the index of the effect to be plotted or use "all" to plot all effects}

\item{title}{character}

\item{cred.band}{logical, if TRUE, plot credible bands if obj fitted with wavelets regression. Set as TRUE by default}

\item{lfsr.curve}{logical, if TRUE, plot estimated lfsr of the effect at each base pair  if obj fitted with HMM regression. Set as TRUE by default}

\item{size_line}{numeric, width of the plotted lines}

\item{\dots}{Other arguments..}
}
\description{
Plot specific effect from susiF object
}
