% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_susiF_effect}
\alias{plot_susiF_effect}
\title{Plot specific effect from susiF object}
\usage{
plot_susiF_effect(
  obj,
  effect = "all",
  title = "",
  cred.band = TRUE,
  lfsr.curve = TRUE,
  linewidth = 0.5,
  font_size = 10,
  ...
)
}
\arguments{
\item{obj}{Output of the susiF function.}

\item{effect}{The indices of the effects to be plotted, or use
"all" to plot all effects.}

\item{title}{The title of the plot.}

\item{cred.band}{logical, if TRUE, plot credible bands if obj
fitted with wavelets regression.}

\item{lfsr.curve}{logical, if TRUE, plot estimated lfsr of the
effect at each base pair if obj fitted with HMM regression.}

\item{\dots}{Other arguments.}

\item{size_line}{numeric, width of the plotted lines}
}
\description{
Plot specific effect from susiF object
}
