% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_susiF_pip}
\alias{plot_susiF_pip}
\title{Plot posterior inclusion probabilities from susiF object}
\usage{
plot_susiF_pip(
  obj,
  title = "",
  point_size = 2,
  pos_SNP,
  point_shape,
  font_size = 10
)
}
\arguments{
\item{point_size}{numeric, size of the point}

\item{pos_SNP}{vector, containing the base pair of the SNPs}

\item{point_shape}{vector, containing the shape of dots}

\item{font_size}{numeric, the size of the font}
}
\description{
Plot posterior inclusion probabilities from susiF object
}
