% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_poisson_mean_routines.R
\name{pois_mean_GG}
\alias{pois_mean_GG}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
pois_mean_GG(
  x,
  s = NULL,
  prior_mean = NULL,
  prior_var = NULL,
  optim_method = "L-BFGS-B",
  maxiter = 1000,
  tol = 1e-05
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{prior_mean}{prior mean}

\item{prior_var}{prior variance}

\item{optim_method}{optimization method in `optim` function}

\item{maxiter}{max number of iterations}

\item{tol}{tolerance for stopping the updates}

\item{w}{prior weights}
}
\value{
a list of
 \item{posteriorMean:}{posterior mean}
 \item{posteriorVar:}{posterior variance}
 \item{obj_value:}{objective function values}
 \item{prior_mean:}{prior mean}
 \item{prior_var:}{prior variance}
 @example
 n=300
x = rpois(n,exp(2*sin(1:n/20)))
naive=pois_mean_GG(x)
prior_base= pois_mean_GG(x,
                        prior_mean = 2*sin(1:n/20),
                        prior_var=rep(1, length(n)))
plot(prior_base$posterior$posteriorMean_latent, col="green", type="l")
lines(naive$posterior$posteriorMean_latent)
points(2*sin(1:n/20), pch=19)
}
\description{
Gaussian prior, Gaussian posterior in Poisson mean problem.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(\beta,\sigma^2).}
}
