% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_poisson_mean_routines.R
\name{pois_mean_GG_opt_obj}
\alias{pois_mean_GG_opt_obj}
\title{calculate objective function}
\usage{
pois_mean_GG_opt_obj(theta, ..., x, s, beta, sigma2, n)
}
\description{
calculate objective function
}
