% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_poisson_mean_routines.R
\name{pois_mean_GP}
\alias{pois_mean_GP}
\title{Solve Gaussian approximation to Poisson mean problem}
\usage{
pois_mean_GP(
  x,
  s = NULL,
  prior_mean = NULL,
  prior_var = NULL,
  optim_method = "L-BFGS-B",
  maxiter = 1000,
  tol = 1e-05
)
}
\arguments{
\item{x}{data vector}

\item{s}{scaling vector}

\item{prior_mean}{prior mean}

\item{prior_var}{prior variance}

\item{optim_method}{optimization method in `optim` function}

\item{maxiter}{max number of iterations}

\item{tol}{tolerance for stopping the updates}

\item{w}{prior weights}
}
\value{
a list of
 \item{posteriorMean:}{posterior mean}
 \item{posteriorVar:}{posterior variance}
 \item{obj_value:}{objective function values}
 \item{prior_mean:}{prior mean}
 \item{prior_var:}{prior variance}
 @example
 n = 10000
 mu = rnorm(n)
 x = rpois(n,exp(mu))
 pois_mean_GG(x)
}
\description{
Gaussian prior, Gaussian posterior in Poisson mean problem.
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(\beta,\sigma^2).}
}
