% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_poisson_mean_routines.R
\name{pois_smooth_split}
\alias{pois_smooth_split}
\title{Smooth over-dispersed Poisson sequence}
\usage{
pois_smooth_split(
  x,
  s = NULL,
  Eb_init = NULL,
  sigma2_init = NULL,
  est_sigma2 = TRUE,
  maxiter = 100,
  tol = 1e-05,
  filter.number = 1,
  family = "DaubExPhase",
  verbose = FALSE,
  printevery = 10,
  ebnm_params = list(mode = 0),
  optim_method = "L-BFGS-B"
)
}
\arguments{
\item{x}{data vector}

\item{Eb_init, sigma2_init}{initial values of smooth mean and nugget effect.}

\item{maxiter, tol}{max iteration and tolerance for stopping it.}
}
\description{
Smooth over-dispersed Poisson sequence
}
\details{
The problem is
\deqn{x_i\sim Poisson(\exp(\mu_i)),}
\deqn{\mu_i\sim N(b_i,\sigma^2),}
\deqn{\b_i\sim g(.).}
}
\examples{
set.seed(12345)
n=2^9
sigma=0.5
mu=c(rep(0.3,n/4), rep(3, n/4), rep(10, n/4), rep(0.3, n/4))
x = rpois(n,exp(log(mu)+rnorm(n,sd=sigma)))
fit = pois_smooth_split(x,maxiter=30)
plot(x,col='grey80')
lines(exp(fit$Eb))
fit$sigma2
plot(fit$obj)
}
