% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{post_mat_mean}
\alias{post_mat_mean}
\alias{post_mat_mean.mixture_normal}
\alias{post_mat_mean.mixture_normal_per_scale}
\title{Compute posterior mean under mixture normal prior}
\usage{
post_mat_mean(G_prior, Bhat, Shat, lBF, lowc_wc, indx_lst, e, ...)

\method{post_mat_mean}{mixture_normal}(
  G_prior,
  Bhat,
  Shat,
  lBF = NULL,
  lowc_wc,
  indx_lst,
  e = 0.001,
  ...
)

\method{post_mat_mean}{mixture_normal_per_scale}(
  G_prior,
  Bhat,
  Shat,
  lBF = NULL,
  lowc_wc,
  indx_lst,
  e = 0.001,
  ...
)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{lBF}{log BF}

\item{lowc_wc}{wavelet coefficient with low count to be discarded}

\item{indx_lst}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class mixture_normal_per_scale}

\item{e}{threshold value to avoid computing posterior that have low alpha value. Set it to 0 to compute the entire posterior. default value is 0.001}

\item{\dots}{Other arguments.}
}
\value{
pxJ matrix of posterior mean
}
\description{
Add description here.
}
\keyword{internal}
