% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{post_mat_sd.mixture_normal}
\alias{post_mat_sd.mixture_normal}
\title{Compute posterior standard deviation under mixture normal prior}
\usage{
\method{post_mat_sd}{mixture_normal}(G_prior, Bhat, Shat, ...)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{indx_list}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class mixture_normal_per_scale}
}
\value{
pxJ matrix of posterior standard deviation
}
\description{
Add description here.
}
