% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{post_mean.mixture_normal_per_scale}
\alias{post_mean.mixture_normal_per_scale}
\title{Compute posterior mean under mixture normal per scale prior}
\usage{
\method{post_mean}{mixture_normal_per_scale}(G_prior, Bhat, Shat, indx_lst, ...)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{indx_list}{list generated by gen_wavelet_indx for the given level of resolution, used only with class \class{mixture_normal_per_scale}}
}
\value{
pxJ matrix of posterior mean
}
\description{
Compute posterior mean under mixture normal per scale prior
}
