% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{remap_data}
\alias{remap_data}
\title{Ramp matrix of unevenly space data (or non power of two)}
\usage{
remap_data(Y, pos, verbose = TRUE)
}
\arguments{
\item{Y}{Matrix of observed curves}

\item{pos}{sampling position}

\item{verbose}{logical}
}
\description{
Ramp matrix of unevenly space data (or non power of two)
}
