% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{Shat_recov}
\alias{Shat_recov}
\title{Transform raw regression coefficient standard error to wavelet regression coefficients standard error}
\usage{
Shat_recov(Shat, W1)
}
\arguments{
\item{Shat}{matrix of standard error}

\item{W1}{matrix associated with a wavelet transform}
}
\value{
Matrix of standard error of the wavelet regression coefficient  using wavelet transform from W1
}
\description{
Transform raw regression coefficient standard error to wavelet regression coefficients standard error
}
