% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{Shatwc_recov_vec}
\alias{Shatwc_recov_vec}
\title{Transform a vector of raw regression coefficient standard error to wavelet regression coefficients standard error}
\usage{
Shatwc_recov_vec(sd_vec, W1)
}
\arguments{
\item{sd_vec}{matrix of standard error (must be of power of two)}

\item{W1}{matrix associated with a wavelet transform}
}
\value{
Matrix of standard error of the wavelet regression coefficient  using wavelet transform from W1
}
\description{
Transform a vector of raw regression coefficient standard error to wavelet regression coefficients standard error
}
