% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations_functions.R
\name{simu_test_function}
\alias{simu_test_function}
\title{Simulation using Donoho and Johnstone test functions}
\usage{
simu_test_function(
  N = 50,
  P = 10,
  lev_res = 7,
  rsnr = 2,
  is.plot = TRUE,
  pos1 = 1,
  pos2
)
}
\arguments{
\item{N}{integer number of sample to simulate}

\item{P}{number of covariate}

\item{lev_res}{control length of the generated function (length function = 2^lev_res)}

\item{rsnr}{root signal noise ratio for the noise}

\item{is.plot}{logical if set to TRUE plot underying function}

\item{pos1}{position of the first active covariate}

\item{pos2}{position of the first active covariate (optional)}
}
\description{
Simulation using Donoho and Johnstone test functions
}
