% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_poisson_mean_routines.R
\name{smash_dwt}
\alias{smash_dwt}
\title{Empirical Bayes wavelet smoothing via DWT}
\usage{
smash_dwt(
  x,
  sigma,
  filter.number = 1,
  family = "DaubExPhase",
  ebnm_params = list(mode = 0),
  W = NULL
)
}
\arguments{
\item{x}{data}

\item{sigma}{known standard error}

\item{filter.number, family}{wavelet family and filter number as in wavethresh package}

\item{ebnm_params}{a list of `ebnm` parameters}

\item{W}{the dwt matrix for calc posterior variance. Remove the first row which is all 1/sqrt(n)}
}
\value{
a list of
 \item{mu.est:}{posterior mean}
 \item{mu.est.var:}{posterior variance}
 \item{loglik:}{log likelihood}
 \item{dKL:}{KL divergence between g(the prior) and q(the posterior)}
}
\description{
Smooth homogeneous Gaussian data.
}
