% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{smash_regression}
\alias{smash_regression}
\alias{smash_regression.susiF}
\title{Compute refined estimate using translation invariant wavelet transform}
\usage{
smash_regression(obj, Y, X, verbose, filter.number, family, alpha, ...)

\method{smash_regression}{susiF}(
  obj,
  Y,
  X,
  verbose = TRUE,
  filter.number = 10,
  family = "DaubLeAsymm",
  alpha = 0.99,
  ...
)
}
\arguments{
\item{obj}{a susiF object}

\item{Y}{matrix of responses}

\item{X}{matrix containing the covariates}

\item{verbose}{logical}

\item{filter.number}{see wd description in wavethresh package description}

\item{family}{see wd description in wavethresh package description}

\item{alpha}{required confidence level}

\item{\dots}{Other arguments.}
}
\description{
e Compute refined estimate using translation invariant wavelet transform
}
