% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiF_ss.R
\name{susiF_ss}
\alias{susiF_ss}
\title{Sum of Single Functions}
\usage{
susiF_ss(
  Bhat,
  Shat,
  R,
  n,
  var_y,
  XtX,
  Xty,
  yty,
  L = 2,
  pos = NULL,
  prior = "mixture_normal_per_scale",
  verbose = TRUE,
  plot_out = TRUE,
  maxit = 100,
  tol = 0.001,
  cov_lev = 0.95
)
}
\arguments{
\item{Bhat}{A p by t matrix of estimated effects.}

\item{Shat}{A p by t matrix of standard errors.}

\item{R}{A p by p correlation matrix. It should be estimated from
  the same samples used to compute \code{bhat} and \code{shat}. Using
  an out-of-sample matrix may produce unreliable results.

 @param n The sample size.}

\item{var_y}{a T vector of The sample variance of Y at each time point , defined as \eqn{y'y/(n-1)}.
When the sample variance cannot be provided, the coefficients
(returned from \code{coef}) are computed on the "standardized" X, y
scale.}

\item{XtX}{A p by p matrix \eqn{X'X} in which the columns of X
are centered to have mean zero.}

\item{Xty}{A p by T matrix \eqn{X'y} in which y and the columns of X are
centered to have mean zero.}

\item{yty}{A T by T scalar \eqn{y'y} in which y is centered to have mean
zero.}

\item{pos}{vector of length J, corresponding to position/time pf
the observed column in Y, if missing suppose that the observation
are evenly spaced}

\item{prior}{specify the prior used in susif. Three choice are
available "normal", "mixture_normal", "mixture_normal_per_scale"}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
and a summary of the optimization settings, are printed to the
console.}

\item{plot_out}{If \code{plot_out = TRUE}, the algorithm's progress,
and a summary of the optimization settings, are ploted.}

\item{maxit}{Maximum number of IBSS iterations to perform.}

\item{tol}{A small, non-negative number specifying the convergence
tolerance for the IBSS fitting procedure. The fitting procedure
will halt when the difference in the variational lower bound, or
\dQuote{ELBO} (the objective function to be maximized), is less
than \code{tol}.}

\item{cov_lev}{numeric between 0 and 1, corresponding to the
expected level of coverage of the cs if not specified set to 0.95}
}
\description{
Implementation of the SuSiF method
}
\details{
tbd
}
