% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{test_stop_cond}
\alias{test_stop_cond}
\alias{test_stop_cond.susiF}
\title{Check tolerance for stopping criterion}
\usage{
test_stop_cond(susiF.obj, check, cal_obj, Y, X, D, C, indx_lst, ...)

\method{test_stop_cond}{susiF}(susiF.obj, check, cal_obj, Y, X, D, C, indx_lst)
}
\description{
Check tolerance for stopping criterion
}
