% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{TI_regression}
\alias{TI_regression}
\alias{TI_regression.susiF}
\title{Compute refined estimate using translation invariant wavelet transform}
\usage{
TI_regression(obj, Y, X, verbose, filter.number, family, alpha, ...)

\method{TI_regression}{susiF}(
  obj,
  Y,
  X,
  verbose = TRUE,
  filter.number = 1,
  family = "DaubExPhase",
  alpha = 0.05,
  n_iter = 5,
  ...
)
}
\arguments{
\item{obj}{a susiF object}

\item{Y}{matrix of responses}

\item{X}{matrix containing the covariates}

\item{verbose}{logical}

\item{filter.number}{see wd description in wavethresh package description}

\item{family}{see wd description in wavethresh package description}

\item{alpha}{required confidence level}

\item{\dots}{Other arguments.}

\item{n_iter}{numeric number of iteration}
}
\description{
Compute refined estimate using translation invariant wavelet transform
}
