% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{trans_sum_stat_wreg}
\alias{trans_sum_stat_wreg}
\title{Transform raw regression coefficient to wavelet regression coefficients}
\usage{
trans_sum_stat_wreg(Bhat, Shat)
}
\arguments{
\item{Bhat}{matrix of regression coefficient}

\item{Shat}{matrix of standard error}
}
\value{
list of two

\item{Bhat}{ matrix pxJ wavelet  regression coefficient}
\item{Shat}{ matrix pxJ  wavelet  regression coefficient standard error}
}
\description{
Transform raw regression coefficient to wavelet regression coefficients
}
