% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_alpha}
\alias{update_alpha}
\alias{update_alpha.susiF}
\title{Update alpha   susiF mixture proportion of effect l}
\usage{
update_alpha(susiF.obj, l, alpha, ...)

\method{update_alpha}{susiF}(susiF.obj, l, alpha, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by init_susiF_obj function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{alpha}{vector of p alpha values summing up to one}
}
\value{
susiF object
}
\description{
Update alpha   susiF mixture proportion of effect l
}
\keyword{internal}
