% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{update_alpha.susiF}
\alias{update_alpha.susiF}
\title{Update alpha   susif mixture proportion of effect l}
\usage{
\method{update_alpha}{susiF}(susiF.obj, l, alpha)
}
\arguments{
\item{susiF.obj}{a susisf object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{alpha}{vector of p alpha values summing up to one}
}
\value{
susiF object
}
\description{
Update alpha   susif mixture proportion of effect l
}
