% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_cal_fit_func.susiF}
\alias{update_cal_fit_func.susiF}
\title{Update susiF by computing posterior curves}
\usage{
\method{update_cal_fit_func}{susiF}(susiF.obj, indx_lst)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelet which assume that J is of the form J^2. If J not a power of 2, susif internally remaps the data into grid of length 2^J}

\item{X}{matrix of size n by p
#'@param indx_lst list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
susiF object
}
\description{
Update susiF by computing posterior curves
}
