% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_cal_lfsr_func}
\alias{update_cal_lfsr_func}
\alias{update_cal_lfsr_func.susiF}
\title{Update susiF by computing posterior curves using wavelet coefficient with a low lfsr}
\usage{
update_cal_lfsr_func(susiF.obj, lfsr_curve, indx_lst, ...)

\method{update_cal_lfsr_func}{susiF}(susiF.obj, lfsr_curve, indx_lst, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by init_susiF_obj function}

\item{lfsr_curve}{Maximum local false sign rate of the wavelet coefficients used to reconstruct lfsr_curves (see output)}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
susiF object
}
\description{
Update susiF by computing posterior curves using wavelet coefficient with a low lfsr
}
