% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_cal_pip}
\alias{update_cal_pip}
\alias{update_cal_pip.susiF}
\title{Update susiF by computing PiP}
\usage{
update_cal_pip(obj, ...)

\method{update_cal_pip}{susiF}(obj, ...)
}
\arguments{
\item{obj}{a susiF object defined by  init_susiF_obj  function}
}
\value{
susiF object
}
\description{
Update susiF by computing PiP
}
\keyword{internal}
