% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{update_cal_pip.susiF}
\alias{update_cal_pip.susiF}
\title{Update susiF by computing PiP}
\usage{
\method{update_cal_pip}{susiF}(susiF.obj)
}
\arguments{
\item{susiF.obj}{a susisf object defined by init_susiF_obj function}
}
\value{
susiF object
}
\description{
Update susiF by computing PiP
}
