% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_ELBO}
\alias{update_ELBO}
\alias{update_ELBO.susiF}
\title{Update susiF log Bayes factor}
\usage{
update_ELBO(susiF.obj, ELBO, ...)

\method{update_ELBO}{susiF}(susiF.obj, ELBO, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{ELBO}{new ELBO value}
}
\value{
susiF object
}
\description{
Update susiF log Bayes factor
}
\keyword{internal}
