% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{update_expected_residual}
\alias{update_expected_residual}
\alias{update_expected_residual.susiF_ss}
\title{Update expected residuals for susiF_ss.obj}
\usage{
update_expected_residual(susiF_ss.obj, data, l)

\method{update_expected_residual}{susiF_ss}(susiF_ss.obj, data, l)
}
\arguments{
\item{susiF_ss.obj}{susiF_ss.obj object}

\item{data}{an object of the class suff_stat define by function \code{\link{make_data_suff_stat}}}

\item{l, }{optional effect to update}
}
\value{
an updated suff_stat object
}
\description{
Update expected residuals for susiF_ss.obj for effect l
}
