% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{update_KL}
\alias{update_KL}
\alias{update_KL.susiF}
\title{Compute KL divergence effect l}
\usage{
update_KL(susiF.obj, X, D, C, indx_lst, ...)

\method{update_KL}{susiF}(susiF.obj, Y, X, D, C, indx_lst, ...)
}
\arguments{
\item{X}{matrix of covariate}

\item{D}{matrix of wavelet D coefficients from the original input data (Y)}

\item{C}{vector of wavelet scaling coefficient from the original input data (Y)}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{Y}{wavelet transformed phenotype}
}
\value{
susiF object
}
\description{
Compute KL divergence effect l
}
