% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{update_lBF}
\alias{update_lBF}
\alias{update_lBF.susiF}
\alias{update_lBF.susiF_ss}
\title{Update susiF log Bayes factor}
\usage{
update_lBF(susiF.obj, l, lBF, ...)

\method{update_lBF}{susiF}(susiF.obj, l, lBF, ...)

\method{update_lBF}{susiF_ss}(susiF_ss.obj, l, lBF, ...)
}
\arguments{
\item{susiF.obj}{a susisf object defined by \code{\link{init_susiF_obj}} function}

\item{l}{effect to update}

\item{lBF}{vector of length p, containning the updated log Bayes factors}
}
\value{
susiF object
}
\description{
Update susiF log Bayes factor
}
