% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{update_pi.susiF}
\alias{update_pi.susiF}
\title{Update mixture proportion of susif mixture proprotion of effect l}
\usage{
\method{update_pi}{susiF}(susiF.obj, l, tpi)
}
\arguments{
\item{susiF.obj}{a susisf object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{tpi}{an object of the class "pi_mixture_normal" or "pi_mixture_normal_per_scale"}
}
\value{
susiF object
}
\description{
Update mixture proportion of susif mixture proprotion of effect l
}
