% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{update_prior.mixture_normal}
\alias{update_prior.mixture_normal}
\title{Update mixture proportion for mixture normal prior}
\usage{
\method{update_prior}{mixture_normal}(G_prior, tpi)

\method{update_prior}{mixture_normal}(G_prior, tpi)
}
\arguments{
\item{G_prior}{a prior of class "mixture_normal"}

\item{tpi}{a vector of proportion of class"pi_mixture_normal"}
}
\value{
a prior of class "mixture_normal"

a prior of class "mixture_normal"
}
\description{
Add description here.

Add description here.
}
