% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{update_prior.mixture_normal_per_scale}
\alias{update_prior.mixture_normal_per_scale}
\title{Update mixture proportion for mixture normal per scale prior}
\usage{
\method{update_prior}{mixture_normal_per_scale}(G_prior, tpi)
}
\arguments{
\item{tpi}{a vector of proportion of class"pi_mixture_normal_per_scale"}

\item{G}{a prior of class "mixture_normal_per_scale"}
}
\value{
a prior of class "mixture_normal_per_scale"
}
\description{
Update mixture proportion for mixture normal per scale prior
}
