% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{update_prior}
\alias{update_prior}
\alias{update_prior.mixture_normal}
\alias{update_prior.mixture_normal_per_scale}
\title{Update mixture proportion for mixture normal prior}
\usage{
update_prior(G_prior, tpi, ...)

\method{update_prior}{mixture_normal}(G_prior, tpi, ...)

\method{update_prior}{mixture_normal_per_scale}(G_prior, tpi, ...)
}
\arguments{
\item{G_prior}{a prior of class "mixture_normal" or a prior of class "mixture_normal_per_scale"}

\item{tpi}{a vector of proportion of class "pi_mixture_normal" resp "pi_mixture_normal_per_scale"}

\item{\dots}{Other arguments.}
}
\value{
a prior of class "mixture_normal" or a prior of class "mixture_normal_per_scale"
}
\description{
Add description here.
}
