% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_residual_variance}
\alias{update_residual_variance}
\alias{update_residual_variance.susiF}
\title{Update residual variance}
\usage{
update_residual_variance(susiF.obj, sigma2, ...)

\method{update_residual_variance}{susiF}(susiF.obj, sigma2, ...)
}
\arguments{
\item{susiF.obj}{a susiF object}

\item{sigma2}{the new value for residual variance}
}
\description{
See title
}
\keyword{internal}
