% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_susiF_obj}
\alias{update_susiF_obj}
\alias{update_susiF_obj.susiF}
\title{Update  susiF object using the output of EM_pi}
\usage{
update_susiF_obj(
  susiF.obj,
  l,
  EM_pi,
  Bhat,
  Shat,
  indx_lst,
  lowc_wc = NULL,
  cal_wc_lsfr = FALSE,
  ...
)

\method{update_susiF_obj}{susiF}(
  susiF.obj,
  l,
  EM_pi,
  Bhat,
  Shat,
  indx_lst,
  lowc_wc = NULL,
  cal_wc_lsfr = FALSE,
  ...
)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{EM_pi}{an object of the class "EM_pi" generated by the function \code{\link{EM_pi}}}

\item{Bhat}{matrix of estimated regression coefficients}

\item{Shat}{matrix of estimated standard errors}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{lowc_wc}{list of wavelet coefficients that exhibit too little variance}

\item{cal_wc_lsfr}{logical, set to TRUE if detailed output needed}
}
\value{
susiF object
}
\description{
Update  susiF object using the output of EM_pi
}
