% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_susiF_obj.susiF}
\alias{update_susiF_obj.susiF}
\title{Update  susiF object using the output of EM_pi}
\usage{
\method{update_susiF_obj}{susiF}(susiF.obj, l, EM_pi, Bhat, Shat, indx_lst)
}
\arguments{
\item{susiF.obj}{a susisf object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{EM_pi}{an object of the class "EM_pi" generated by the function EM_pi}
}
\value{
susiF object
}
\description{
Update  susiF object using the output of EM_pi
}
