% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{update_susiF_ss_obj}
\alias{update_susiF_ss_obj}
\title{update susiF_ss object}
\usage{
update_susiF_ss_obj(susiF_ss.obj, l, EM_pi, Bhat, Shat, indx_lst, ...)
}
\arguments{
\item{susiF_ss.obj}{an object of class susiF_ss.}

\item{l}{the effect to update}

\item{EM_pi}{the output of EM_pi}

\item{Bhat}{estimated regression coefficients}

\item{Shat}{estimated standard error}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\description{
update susiF_ss object
}
