% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EB_poisson_mean_routines.R
\name{vga_pois_solver_newton_1iter}
\alias{vga_pois_solver_newton_1iter}
\title{Optimize vga poisson problem 1 iteration.}
\usage{
vga_pois_solver_newton_1iter(m, v, y, s, beta, sigma2)
}
\description{
This function only performs vga for 1 iteration
}
