% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{which_cs_largeBF}
\alias{which_cs_largeBF}
\title{which_cs_largeBF}
\usage{
which_cs_largeBF(obj, l1, l2)
}
\arguments{
\item{l1}{cs1}

\item{l2}{cs2}
}
\description{
which_cs_largeBF
}
\keyword{internal}
