% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_lowcount}
\alias{which_lowcount}
\title{Find wavelet with critically low variance}
\usage{
which_lowcount(Y_f, thresh_lowcount)
}
\arguments{
\item{Y_f}{a matrix of wavelet coefficients (col= wc, row =ind)}

\item{thresh_lowcount}{mad}
}
\description{
Find wavelet with critically low variance
}
\keyword{internal}
