% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{adjust_FM_covariate}
\alias{adjust_FM_covariate}
\title{Adjusting functions for covariates}
\usage{
adjust_FM_covariate(
  Y,
  X,
  pos = NULL,
  Batch = NULL,
  thresh_lowcount = 0,
  verbose
)
}
\arguments{
\item{Y}{observed curves.}

\item{X}{matrix containing the covariates}

\item{pos}{Original position of the column of Y}

\item{thresh_lowcount}{numeric, used to check the wavelet coefficients have
problematic distribution (very low dispersion even after standardization).
Basically check if the median of the absolute value of the distribution of
 a wavelet coefficient is below this threshold, if yes, the algorithm discard
 this wavelet coefficient (setting its estimate effect to 0 and estimate sd to 1).
 Set to 0 by default. It can be useful when analyzing sparse data from sequence
  based assay or small samples.}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
and a summary of the optimization settings are printed to the
console.}
}
\description{
Simple function to adjust the observed curves for covariates (e.g., age, population structure)
}
