% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{cal_Bhat_Shat}
\alias{cal_Bhat_Shat}
\title{Regress Y nxJ on X nxp}
\usage{
cal_Bhat_Shat(Y, X, v1)
}
\arguments{
\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelet which assume that J is of the form J^2. If J not a power of 2, susif internally remaps the data into grid of length 2^J}

\item{X}{matrix of size n by p in}
}
\value{
list of two
\item{Bhat}{ matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j] }
\item{Shat}{ matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j] }
}
\description{
regression coefficients (and sd) of the column wise regression
}
