% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{cal_KL_l}
\alias{cal_KL_l}
\alias{cal_KL_l.susiF}
\alias{cal_KL_l.susiF_ss}
\title{Compute KL divergence effect l}
\usage{
cal_KL_l(susiF.obj, l, Y, X, D, C, indx_lst, ...)

\method{cal_KL_l}{susiF}(susiF.obj, l, X, D, C, indx_lst, ...)

\method{cal_KL_l}{susiF_ss}(susiF_ss.obj, l, data, ...)
}
\arguments{
\item{susiF.obj}{a susisF object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{Y}{wavelet transformed phenotype}

\item{X}{matrix of covariate}

\item{D}{matrix of wavelet D coefficients from the original input data (Y)}

\item{C}{vector of wavelet scaling coefficient from the original input data (Y)}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
The KL divergence for effect l
}
\description{
Compute KL divergence effect l
}
