% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{cal_lik}
\alias{cal_lik}
\title{Compute likelihood for the weighted ash problem}
\usage{
cal_lik(lBF, zeta)
}
\arguments{
\item{lBF}{vector of log Bayes factors}

\item{zeta}{assignement probabilities}
}
\value{
Likelihood value
}
\description{
Compute likelihood for the weighted ash problem
}
