% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{cal_zeta}
\alias{cal_zeta}
\title{Compute assignment probabilities from log Bayes factors}
\usage{
cal_zeta(lBF)
}
\arguments{
\item{lBF}{vector of log Bayes factors}
}
\description{
Compute assignment probabilities from log Bayes factors
}
