% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{check_cs}
\alias{check_cs}
\alias{check_cs.susiF}
\title{Check purity credible sets}
\usage{
check_cs(susiF.obj, min.purity = 0.5, X, ...)

\method{check_cs}{susiF}(susiF.obj, min.purity = 0.5, X, ...)
}
\arguments{
\item{susiF.obj}{a susif object defined by \code{\link{init_susiF_obj}} function}

\item{min.purity}{minimal purity within a CS}

\item{X}{matrix of covariates}
}
\value{
a susif.obj without "dummy" credible s
}
\description{
Check purity credible sets
}
\keyword{internal}
