% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{discard_cs}
\alias{discard_cs}
\alias{discard_cs.susiF}
\title{Discard credible sets}
\usage{
discard_cs(susiF.obj, cs, out_prep, ...)

\method{discard_cs}{susiF}(susiF.obj, cs, out_prep = FALSE, ...)
}
\arguments{
\item{susiF.obj}{a susif object defined by init_susiF_obj function}

\item{cs}{vector of integer containing the credible sets to discard}

\item{out_prep}{logical, if set to true perform cleaning for final output}
}
\value{
a susif.obj without "dummy" credible sets
}
\description{
Discard credible sets
}
\keyword{internal}
