% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBmvFR_workhorse.R
\name{EBmvFR.workhorse}
\alias{EBmvFR.workhorse}
\title{Empirical Bayes multivariate functional regression}
\usage{
EBmvFR.workhorse(
  obj,
  W,
  X,
  tol,
  lowc_wc,
  init_pi0_w,
  control_mixsqp,
  indx_lst,
  nullweight,
  cal_obj,
  verbose,
  maxit,
  max_step_EM
)
}
\arguments{
\item{obj}{an object of class EBmvFR}

\item{W}{a list in which element D contains matrix of wavelet d coefficients and
element C contains the vector of scaling coefficients}

\item{X}{matrix of size n by p contains the covariates}

\item{tol}{a small, non-negative number specifying the convergence
tolerance for the IBSS fitting procedure. The fitting procedure
will halt when the difference in the variational lower bound, or
\dQuote{ELBO} (the objective function to be maximized), is less
than \code{tol}.}

\item{lowc_wc}{list of wavelet coefficients that exhibit too little variance}

\item{init_pi0_w}{starting value of weight on null compoenent in mixsqp
(between 0 and 1)}

\item{control_mixsqp}{list of parameter for mixsqp function see  mixsqp package}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{nullweight}{numeric value for penalizing likelihood at point mass 0
(useful in small sample size)}

\item{cal_obj}{logical if set as TRUE compute ELBO for convergence monitoring}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
and a summary of the optimization settings are printed to the
console.}

\item{maxit}{Maximum number of IBSS iterations.}

\item{max_step_EM}{see susiF function}
}
\description{
Empirical Bayes multivariate functional regression
}
\details{
Empirical Bayes multivariate functional regression
}
