% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{Eloglik}
\alias{Eloglik}
\alias{Eloglik.susiF}
\title{Expected log likelihood for a   susiF   object}
\usage{
Eloglik(susiF.obj, Y, X, ...)

\method{Eloglik}{susiF}(susiF.obj, Y, X, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by  init_susiF_obj  function}

\item{Y}{Matrix of outcomes}

\item{X}{Matrix of covariates}
}
\value{
Expected log likelihood
}
\description{
Expected log likelihood for a   susiF   object
}
\keyword{internal}
