% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_EBmvFR_obj.R,
%   R/operation_on_susiF_obj.R
\name{estimate_residual_variance.EBmvFR}
\alias{estimate_residual_variance.EBmvFR}
\alias{estimate_residual_variance}
\alias{estimate_residual_variance.susiF}
\title{Update residual variance}
\usage{
\method{estimate_residual_variance}{EBmvFR}(obj, Y, X, ...)

estimate_residual_variance(susiF.obj, Y, X, ...)

\method{estimate_residual_variance}{susiF}(susiF.obj, Y, X, ...)
}
\arguments{
\item{Y}{wavelet transformed  functional phenotype, matrix of size N by size J.}

\item{X}{matrix of size N by p}

\item{susiF.obj}{a susiF object defined by init_susiF_obj function}
}
\value{
estimated residual variance
}
\description{
Update residual variance
}
\keyword{internal}
