% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{estimate_residual_variance}
\alias{estimate_residual_variance}
\alias{estimate_residual_variance.susiF}
\title{Update residual variance}
\usage{
estimate_residual_variance(susiF.obj, Y, X, ...)

\method{estimate_residual_variance}{susiF}(susiF.obj, Y, X, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{Y}{wavelet transformed  functional phenotype, matrix of size N by size J.}

\item{X}{matrix of size N by p}
}
\value{
estimated residual variance
}
\description{
Update residual variance
}
