% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{estimate_residual_variance_ss}
\alias{estimate_residual_variance_ss}
\title{Estimate residual variance for summary stat}
\usage{
estimate_residual_variance_ss(susiF_ss.obj, data, ...)
}
\description{
Estimate residual variance for summary stat
}
