% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{gen_wavelet_indx}
\alias{gen_wavelet_indx}
\title{Generate list of index wavelet coefficients}
\usage{
gen_wavelet_indx(lev_res)
}
\arguments{
\item{lev_res}{integer, corresponding to log2 of the signal length. WARNING the ordering change for different values of s.}
}
\description{
Generate list of index corresponding to where the wavelet coefficient of scale s are stored in wd$D.
}
\examples{
library(wavethresh)
tem_func <- rnorm( 2^8)
twav <- wd(tem_func)
indx_lst <- gen_wavelet_indx(8)
plot(accessD(twav,level=6), ( twav$D[unlist(indx_lst[(6+1)])]) )
 #should a straightline
}
\keyword{internal}
