% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{get_alpha.susiF}
\alias{get_alpha.susiF}
\title{Update alpha   susif mixture proportion of effect l}
\usage{
\method{get_alpha}{susiF}(susiF.obj, l)
}
\arguments{
\item{susiF.obj}{a susisf object defined by init_susiF_obj function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{alpha}{vector of p alpha values summing up to one}
}
\value{
susiF object
}
\description{
Update alpha   susif mixture proportion of effect l
}
