% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_ss_obj.R
\name{get_ER2_ss}
\alias{get_ER2_ss}
\title{compute expected sum of squares for sum stat model}
\usage{
get_ER2_ss(susiF_ss.obj, data, ...)
}
\description{
compute expected sum of squares for sum stat model
}
