% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{get_fitted_effect}
\alias{get_fitted_effect}
\alias{get_fitted_effect.susiF}
\title{Access fitted effect l}
\usage{
get_fitted_effect(susiF.obj, l, ...)

\method{get_fitted_effect}{susiF}(susiF.obj, l, ...)
}
\arguments{
\item{susiF.obj}{a fitted objected}

\item{l}{effect of interest}
}
\description{
Access fitted effect l
}
