% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{get_fitted_effect}
\alias{get_fitted_effect}
\alias{get_fitted_effect.susiF}
\title{Access fitted effect l}
\usage{
get_fitted_effect(obj, l, cred_band, alpha, ...)

\method{get_fitted_effect}{susiF}(obj, l = 1, cred_band = FALSE, alpha = 0.01, ...)
}
\arguments{
\item{obj}{A fitted object.}

\item{l}{Effect of interest.}

\item{cred_band}{Logical, default set to `FALSE`. If `TRUE`, also returns the credible band.}

\item{alpha}{Numerical, defined as 1 - conf_level   set to obtain  0.99 confidence level by default.}

\item{...}{Other arguments.}
}
\value{
The fitted effect.
}
\description{
Access fitted effect
}
\details{
Retrieves the estimated effect from a fitted model.
}
