% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{get_G_prior}
\alias{get_G_prior}
\alias{get_G_prior.susiF}
\title{Access susiF internal prior}
\usage{
get_G_prior(susiF.obj, ...)

\method{get_G_prior}{susiF}(susiF.obj, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}
}
\value{
G_prior object
}
\description{
Access susiF internal prior
}
\keyword{internal}
