% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{get_G_prior.susiF}
\alias{get_G_prior.susiF}
\title{Access susif internal prior}
\usage{
\method{get_G_prior}{susiF}(susiF.obj)
}
\arguments{
\item{susiF.obj}{a susisf object defined by init_susiF_bj function}
}
\value{
G_prior object
}
\description{
Access susif internal prior
}
