% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{get_objective}
\alias{get_objective}
\alias{get_objective.susiF}
\title{Get objective function from data and susiF object}
\usage{
get_objective(susiF.obj, Y, X, ...)

\method{get_objective}{susiF}(susiF.obj, Y, X, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{Y}{Matrix of outcomes}

\item{X}{Matrix of covariates}
}
\value{
posterior expected log likelihood for  single function regression of effect l
}
\description{
Get objective function from data and susiF object
}
