% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{get_pi0}
\alias{get_pi0}
\alias{get_pi0.pi_mixture_normal}
\alias{get_pi0.pi_mixture_normal_per_scale}
\title{Get mixture proportion for mixture normal prior}
\usage{
get_pi0(tpi, ...)

\method{get_pi0}{pi_mixture_normal}(tpi, ...)

\method{get_pi0}{pi_mixture_normal_per_scale}(tpi, ...)
}
\arguments{
\item{tpi}{object of class pi_mixture_normal}
}
\value{
numeric between 0 an 1
}
\description{
Add description here.
}
\keyword{internal}
