% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{get_post_F}
\alias{get_post_F}
\alias{get_post_F.susiF}
\alias{get_post_F.susiF_ss}
\title{Compute posterior mean of the fitted effect}
\usage{
get_post_F(susiF.obj, l, ...)

\method{get_post_F}{susiF}(susiF.obj, l, ...)

\method{get_post_F}{susiF_ss}(susiF_ss.obj, l, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{l}{, optional effect to update}
}
\value{
A J by T matrix of posterior wavelet coefficient,
\item{if l  missing}{return sum of the  effect  posterior mean }
\item{if l not missing}{return effect specific posterior mean}
}
\description{
Compute posterior mean of the fitted effect
}
