% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{get_post_F2}
\alias{get_post_F2}
\alias{get_post_F2.susiF}
\alias{get_post_F2.susiF_ss}
\title{Compute posterior second moment}
\usage{
get_post_F2(susiF.obj, l, ...)

\method{get_post_F2}{susiF}(susiF.obj, l, ...)

\method{get_post_F2}{susiF_ss}(susiF.obj, l, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{l}{, optional effect to update}
}
\value{
A J by T matrix of posterior wavelet coefficient,
\item{if l  missing}{return sum of the effects  posterior second moment }
\item{if l not missing}{return effect specific posterior second moment}
}
\description{
Compute posterior second moment
}
