% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{greedy_backfit}
\alias{greedy_backfit}
\alias{greedy_backfit.susiF}
\title{Update  susiF via greedy search or backfit}
\usage{
greedy_backfit(susiF.obj, verbose, cov_lev, X, min.purity, ...)

\method{greedy_backfit}{susiF}(susiF.obj, verbose, cov_lev, X, min.purity, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}
}
\value{
susiF object
}
\description{
Update  susiF via greedy search or backfit
}
