% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{HMM_regression}
\alias{HMM_regression}
\alias{HMM_regression.susiF}
\title{Compute refined estimate using HMM regression}
\usage{
HMM_regression(susiF.obj, Y, X, verbose = TRUE, maxit, ...)

\method{HMM_regression}{susiF}(susiF.obj, Y, X, verbose = TRUE, maxit = 5)
}
\arguments{
\item{susiF.obj}{a susiF object}

\item{Y}{matrix of responses}

\item{X}{matrix containing the covariates}

\item{verbose}{logical
@param maxit  max number of iteration}
}
\description{
Compute refined estimate using HMM regression
}
