% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{init_prior.default}
\alias{init_prior.default}
\title{Initialize the prior}
\usage{
\method{init_prior}{default}(
  Y,
  X,
  prior,
  v1,
  indx_lst,
  lowc_wc,
  control_mixsqp,
  nullweight,
  gridmult = sqrt(2)
)
}
\arguments{
\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelet which assume that J is of the form J^2. If J not a power of 2, susif internally remaps the data into grid of length 2^J}

\item{X}{matrix of size n by p in}

\item{prior}{Three choice are available "normal", "mixture_normal", "mixture_normal_per_scale"}

\item{indx_lst}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class  mixture_normal_per_scale}

\item{lowc_wc}{wavelet coefficient with low count to be discarded}
}
\value{
an object of the class "normal", "mixture_normal" or "mixture_normal_per_scale"
}
\description{
generate list of object corresponding to the parameters of the prior
}
