% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{init_susiF_obj}
\alias{init_susiF_obj}
\title{Initialize a susiF object using regression coefficients}
\usage{
init_susiF_obj(L, G_prior, Y, X)
}
\arguments{
\item{L}{number of non zero coefficients An L-vector containing the indices of the
nonzero coefficients.}

\item{G_prior}{prior object defined by init_prior function}

\item{Y}{Matrix of outcomes}

\item{X}{Matrix of covariate}
}
\value{
A list with elements
}
\description{
Initialize a susiF object using regression coefficients
}
