% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{init_susiF_obj}
\alias{init_susiF_obj}
\title{Initialize a susiF object using regression coefficients}
\usage{
init_susiF_obj(
  L_max,
  G_prior,
  Y,
  X,
  L_start,
  greedy,
  backfit,
  tol_null_prior = 0.001,
  cov_lev = 0.95,
  ...
)
}
\arguments{
\item{L_max}{upper bound on the number of non zero coefficients An L-vector containing the indices of the}

\item{G_prior}{prior object defined by init_prior function}

\item{Y}{Matrix of outcomes}

\item{X}{matrix of covariatess}

\item{L_start}{number of effect to start with}

\item{greedy}{logical, if TRUE allow greedy search}

\item{backfit}{logical, if TRUE allow backfitting}

\item{tol_null_prior}{threshold to consider prior to be null. If the estimated weight on the point mass at zero is larger than 1-tol_null_prior
then set prior weight on point mass to be 1. In the mixture normal this corresponds to removing the effect. In the mixutre per scale prior this corresponds
to setting the prior of a given scale to at point mass at 0.}

\item{cov_lev}{numeric between 0 and 1, corresponding to the
expected level of coverage of the CS if not specified, set to 0.95}

\item{\dots}{Other arguments.}
}
\description{
Initialize a susiF object using regression coefficients
}
\details{
Initialize a susiF object using regression coefficients
}
