% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{init_susiF_obj}
\alias{init_susiF_obj}
\title{Initialize a susiF object using regression coefficients}
\usage{
init_susiF_obj(L_max, G_prior, Y, X, L_start, greedy, backfit, ...)
}
\arguments{
\item{L_max}{upper bound on the number of non zero coefficients An L-vector containing the indices of the
nonzero coefficients.}

\item{G_prior}{prior object defined by init_prior function}

\item{Y}{Matrix of outcomes}

\item{X}{Matrix of covariates}

\item{L_start}{number of effect to start with}

\item{greedy}{logical, if TRUE allow greedy search}

\item{backfit}{logical, if TRUE allow backfitting}
}
\value{
A list with the following elements
\item{fitted_wc}{ list of length L, each element contains the fitted wavelet coefficients of effect l}
\item{fitted_wc2}{list of length L, each element contains the variance of the fitted wavelet coefficients of effect l}
\item{alpha_hist}{ history of the fitted alpha value}
\item{N}{ number of indidivual in the study}
\item{sigma2}{residual variance}
\item{n_wac}{number of wavelet coefficients}
\item{ind_fitted_func}{fitted curves of each individual }
\item{cs}{credible set}
\item{pip}{Posterior inclusion probabilites}
\item{G_prior}{a G_prior of the same class as the input G_prior, used for internal calculation}
\item{lBF}{ log Bayes factor for the different effect}
\item{KL}{ the KL divergence for the different effect}
\item{ELBO}{ The evidence lower bound}
\item{lfsr_wc}{Local fasle sign rate of the fitted wavelet coefficients}
}
\description{
Initialize a susiF object using regression coefficients
}
