% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{log_BF}
\alias{log_BF}
\alias{log_BF.mixture_normal}
\alias{log_BF.mixture_normal_per_scale}
\title{Compute Log-Bayes Factor}
\usage{
log_BF(G_prior, Bhat, Shat, ...)

\method{log_BF}{mixture_normal}(G_prior, Bhat, Shat, ...)

\method{log_BF}{mixture_normal_per_scale}(G_prior, Bhat, Shat, indx_lst, ...)
}
\arguments{
\item{G_prior}{Mixture normal prior.}

\item{Bhat}{p x J matrix of regression coefficients;
\code{Bhat[j,t]} gives regression coefficient of \code{Y[,t]} on
\code{X[,j]}.}

\item{Shat}{p x J matrix of standard errors; \code{Shat[j,t]}
gives standard error of the regression coefficient of
\code{Y[,t]} on {X[,j]}.}

\item{indx_list}{List generated by \code{\link{gen_wavelet_indx}}
for the given level of resolution, used only with class
\dQuote{mixture_normal_per_scale}}
}
\value{
The log-Bayes factor for each covariate.
}
\description{
Compute Log-Bayes Factor
}
