% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{loglik_SFR}
\alias{loglik_SFR}
\alias{loglik_SFR.susiF}
\title{Compute log likelihood of single function regression of effect l}
\usage{
loglik_SFR(obj, l, ...)

\method{loglik_SFR}{susiF}(obj, l, Y, X, indx_lst, ...)
}
\arguments{
\item{obj}{a susiF object defined by  init_susiF_obj  function}

\item{l}{effect to update}

\item{Y}{wavelet transformed phenotype}

\item{X}{Matrix of covariates}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
The log-likelihood, \eqn{\log p(Y | X, V)}, where V is the prior parameters
}
\description{
Compute log likelihood of single function regression of effect l
}
\keyword{internal}
