% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{loglik_SFR_post}
\alias{loglik_SFR_post}
\alias{loglik_SFR_post.susiF}
\title{Compute posterior expected loglikelihood for single function}
\usage{
loglik_SFR_post(obj, l, ...)

\method{loglik_SFR_post}{susiF}(obj, l, Y, X, ...)
}
\arguments{
\item{obj}{susiF object, for example created by calling
init_susiF_obj .}

\item{l}{Index of effect to update.}

\item{\dots}{Other arguments.}

\item{Y}{Wavelet-transformed phenotype.}

\item{X}{Matrix of covariates.}
}
\value{
Posterior expected log-likelihood for the single-function
}
\description{
Compute posterior expected loglikelihood for single function
}
\keyword{internal}
