% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{m_step}
\alias{m_step}
\alias{m_step.lik_mixture_normal}
\alias{m_step.lik_mixture_normal_per_scale}
\title{Compute M step in the weighted ash problem for different prior}
\usage{
m_step(L, zeta, indx_lst, init_pi0_w, ...)

\method{m_step}{lik_mixture_normal}(L, zeta, indx_lst, init_pi0_w, ...)

\method{m_step}{lik_mixture_normal_per_scale}(L, zeta, indx_lst, init_pi0_w, ...)
}
\arguments{
\item{L}{output of \code{\link{L_mixsqp}} function}

\item{zeta}{assignment probabilities for each covariate}

\item{init_pi0_w}{starting value of weight on null compoenent in mixsqp}

\item{indx_list}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class  mixture_normal_per_scale}
}
\value{
a vector of proportion (class pi_mixture_normal)
}
\description{
Compute M step in the weighted ash problem for different prior
}
