% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{m_step}
\alias{m_step}
\alias{m_step.lik_mixture_normal}
\alias{m_step.lik_mixture_normal_per_scale}
\title{Compute M step in the weighted ash problem for different prior}
\usage{
m_step(L, zeta, indx_lst, init_pi0_w, control_mixsqp, ...)

\method{m_step}{lik_mixture_normal}(L, zeta, indx_lst, init_pi0_w, control_mixsqp, nullweight, ...)

\method{m_step}{lik_mixture_normal_per_scale}(L, zeta, indx_lst, init_pi0_w = 1, control_mixsqp, nullweight, ...)
}
\arguments{
\item{L}{output of  L_mixsqp  function}

\item{zeta}{assignment probabilities for each covariate}

\item{indx_lst}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class  mixture_normal_per_scale}

\item{init_pi0_w}{starting value of weight on null compoenent in mixsqp}

\item{control_mixsqp}{list of parameter for mixsqp function see mixsqp package}

\item{nullweight}{numeric value for penalizing likelihood at point mass 0 (should be between 0 and 1)
(usefull in small sample size)}
}
\value{
a vector of proportion (class pi_mixture_normal)
}
\description{
Compute M step in the weighted ash problem for different prior
}
