% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{make_data_suff_stat}
\alias{make_data_suff_stat}
\title{Define sufficient stat data format}
\usage{
make_data_suff_stat(Bhat, Shat, R, N, var_y, XtX, Xty, yty, wav_trans = FALSE)
}
\arguments{
\item{Bhat}{A p by t matrix of estimated effects.}

\item{Shat}{A p by t matrix of standard errors.}

\item{R}{A p by p correlation matrix. It should be estimated from
the same samples used to compute \code{Bhat} and \code{Shat}. Using
an out-of-sample matrix may produce unreliable results.}

\item{N}{The sample size.}

\item{var_y}{a T vector of The sample variance of Y at each time point , defined as \eqn{y'y/(n-1)}.
When the sample variance cannot be provided, the coefficients
(returned from \code{coef}) are computed on the "standardized" X, y
scale.}

\item{XtX}{A p by p matrix \eqn{X'X} in which the columns of X
are centered to have mean zero.}

\item{Xty}{A p by T matrix \eqn{X'y} in which y and the columns of X are
centered to have mean zero.}

\item{yty}{A T by T scalar \eqn{y'y} in which y is centered to have mean
zero.}

\item{wav_trans}{logical, if true the algorithm will consider that the summary statistics based on wavelet transformed data (\code{Bhat} and \code{Shat}).
if False, the algorithm will rescale \code{Bhat} and \code{Shat} to obtain summary statistics from wavelet regression. Default set as FALSE .}
}
\description{
Define sufficient stat data format
}
