% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{merge_effect}
\alias{merge_effect}
\alias{merge_effect.susiF}
\title{Merging effect function}
\usage{
merge_effect(susiF.obj, tl, ...)

\method{merge_effect}{susiF}(susiF.obj, tl, discard = TRUE, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{tl}{see  \code{\link{greedy_backfit}}}

\item{discard}{logical, if set to TRUE allow discarding redundant effect}
}
\value{
a susiF object
}
\description{
Merging effect function
}
