% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{out_prep}
\alias{out_prep}
\alias{out_prep.susiF}
\title{Preparing output of main susiF function}
\usage{
out_prep(susiF.obj, Y, X, indx_lst, filter.cs, lfsr_curve, outing_grid, ...)

\method{out_prep}{susiF}(susiF.obj, Y, X, indx_lst, filter.cs, lfsr_curve, outing_grid, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelets that assume that J is of the form J^2. If J is not a power of 2, susiF internally remaps the data into a grid of length 2^J}

\item{X}{matrix of size N by p}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{filter.cs}{logical, if TRUE filter the credible set (removing low purity cs and cs with estimated prior equal to 0)}

\item{lfsr_curve}{Maximum local false sign rate of the wavelet coefficients used to reconstruct lfsr_curves (see output)}

\item{outing_grid}{grid use to fit fsusie}
}
\value{
susiF object
}
\description{
Preparing output of main susiF function
}
\keyword{internal}
