% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{plot_effect}
\alias{plot_effect}
\title{Plot susiF object}
\usage{
plot_effect(susiF.obj, L, size_line = 2, ...)
}
\arguments{
\item{susiF.obj}{output of the susiF function}

\item{L}{maximum number of effect to display}
}
\description{
Plot susiF object
}
