% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{plot_susiF}
\alias{plot_susiF}
\title{Plot susiF object}
\usage{
plot_susiF(susiF.obj, cred.band = TRUE, effect, ...)
}
\arguments{
\item{susiF.obj}{output of the susiF function}

\item{cred.band}{logical if set as true plot credible bands. Set as TRUE by default}

\item{effect}{numerical if specified plot on effect in particular}
}
\description{
Plot susiF object
}
