% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_susiF}
\alias{plot_susiF}
\title{Plot susiF object}
\usage{
plot_susiF(
  obj,
  title = "",
  effect = "all",
  cred.band = TRUE,
  lfsr.curve = TRUE,
  linewidth = 1.5,
  point_size = 2,
  pos_SNP,
  point_shape,
  pip_only = FALSE
)
}
\arguments{
\item{obj}{output of the susiF function}

\item{title}{character}

\item{effect}{the index of the effect to be plotted or use "all" to plot all effects}

\item{cred.band}{logical, if TRUE, plot credible bands if obj fitted with wavelets regression. Set as TRUE by default}

\item{lfsr.curve}{logical, if TRUE, plot estimated lfsr of the effect at each base pair  if obj fitted with HMM regression. Set as TRUE by default}

\item{linewidth}{numeric, width of the plotted lines}

\item{point_size}{numeric, size of the point}

\item{pos_SNP}{vector, containing the base pair of the SNPs}

\item{point_shape}{vector, containing the shape of dots}

\item{pip_only}{logical, if TRUE only ouput the PIP plot}
}
\description{
Plot susiF object
}
