% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{post_mat_mean}
\alias{post_mat_mean}
\alias{post_mat_mean.mixture_normal}
\alias{post_mat_mean.mixture_normal_per_scale}
\title{Compute posterior mean under mixture normal prior}
\usage{
post_mat_mean(G_prior, Bhat, Shat, lowc_wc, ...)

\method{post_mat_mean}{mixture_normal}(G_prior, Bhat, Shat, lowc_wc, indx_lst, ...)

\method{post_mat_mean}{mixture_normal_per_scale}(G_prior, Bhat, Shat, lowc_wc, indx_lst, ...)
}
\arguments{
\item{G_prior}{mixture normal prior}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{lowc_wc}{wavelet coefficient with low count to be discarded}

\item{\dots}{Other arguments.}

\item{indx_lst}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class mixture_normal_per_scale}
}
\value{
pxJ matrix of posterior mean
}
\description{
Add description here.
}
\keyword{internal}
