% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_utils.R
\name{remap_data}
\alias{remap_data}
\title{Ramp matrix of unevenly space data (or non power of two)}
\usage{
remap_data(Y, pos, verbose = TRUE, max_scale = 10)
}
\arguments{
\item{Y}{Matrix of observed curves}

\item{pos}{sampling position}

\item{verbose}{logical}

\item{max_scale}{numeric, define the maximum of wavelet coefficients used in the analysis (2^max_scale).
Set 10 true by default.}
}
\description{
Ramp matrix of unevenly space data (or non power of two)
}
