% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{scale_m_step}
\alias{scale_m_step}
\title{Subroutine to compute M step in the weighted ash problem for normal mixture prior per scale at a given scale s}
\usage{
scale_m_step(L, s, zeta, indx_lst, init_pi0_w = 0.5, control_mixsqp, ...)
}
\arguments{
\item{L}{output of the L_mixsqp.mixture_normal_per_scale function}

\item{s}{scale}

\item{zeta}{assignment probabilities for each covariate}

\item{init_pi0_w}{starting value of weight on null compoenent in mixsqp}

\item{control_mixsqp}{list of parameter for mixsqp function see\link{\code{mixsqp}}}

\item{indx_list}{list generated by \code{\link{gen_wavelet_indx}} for the given level of resolution, used only with class  mixture_normal_per_scale}
}
\value{
a vector of proportion for the scale s
}
\description{
Subroutine to compute M step in the weighted ash
}
