% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiF_workhorse.R
\name{susiF.workhorse}
\alias{susiF.workhorse}
\title{Sum of Single Function}
\usage{
susiF.workhorse(
  obj,
  W,
  X,
  tol,
  init_pi0_w,
  control_mixsqp,
  indx_lst,
  lowc_wc,
  nullweight,
  cal_obj,
  verbose,
  cov_lev,
  min_purity,
  maxit,
  tt,
  parallel = FALSE,
  max_SNP_EM = 500,
  max_step_EM = 1,
  cor_small = FALSE,
  is.pois = FALSE,
  e = 0.001,
  estimate_sigma2 = TRUE
)
}
\arguments{
\item{obj}{an object of class susiF}

\item{W}{a list in which element D contains matrix of wavelet d coefficients and
element C contains the vector of scaling coefficients}

\item{X}{matrix of size n by p contains the covariates}

\item{tol}{a small, non-negative number specifying the convergence
tolerance for the IBSS fitting procedure. The fitting procedure
will halt when the difference in the variational lower bound, or
\dQuote{ELBO} (the objective function to be maximized), is less
than \code{tol}.}

\item{init_pi0_w}{starting value of weight on null compoenent in mixsqp
(between 0 and 1)}

\item{control_mixsqp}{list of parameter for mixsqp function see  mixsqp package}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{lowc_wc}{list of wavelet coefficients that exhibit too little variance}

\item{nullweight}{numeric value for penalizing likelihood at point mass 0 (should be between 0 and 1)
(usefull in small sample size)}

\item{cal_obj}{logical if set as TRUE compute ELBO for convergence monitoring}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
and a summary of the optimization settings are printed to the
console.}

\item{cov_lev}{numeric between 0 and 1, corresponding to the
expected level of coverage of the cs if not specified set to 0.95}

\item{min_purity}{minimum purity for estimated credible sets}

\item{maxit}{Maximum number of IBSS iterations.}

\item{tt}{output of the cal_Bhat_Shat function, if provided allow skipping the first step of the IBSS}

\item{parallel}{if true use parallel computation}

\item{max_SNP_EM}{check susiF  description}

\item{max_step_EM}{max_step_EM}

\item{cor_small}{check susiF  description}

\item{is.pois}{check susiF  description}

\item{e}{threshold value to avoid computing posterior that have low alpha value. Set it to 0 to compute the entire posterio. default value is 0.001}

\item{estimate_sigma2}{Logical; if \code{FALSE}, keep \code{obj$sigma2} fixed during
all IBSS updates.}
}
\description{
Implementation of the SuSiF method
}
\details{
Implementation of the SuSiF method
}
