% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiF_workhorse.R
\name{susiF.workhorse}
\alias{susiF.workhorse}
\title{Sum of Single Function}
\usage{
susiF.workhorse(
  susiF.obj,
  W,
  X,
  tol,
  low_wc,
  init_pi0_w,
  control_mixsqp,
  indx_lst,
  lowc_wc,
  nullweight,
  cal_obj,
  verbose,
  cov_lev,
  min.purity,
  maxit,
  tt,
  parallel = FALSE
)
}
\arguments{
\item{susiF.obj}{an object of class susiF}

\item{W}{a list in which element D contains matrix of wavelet d coefficients and
element C contains the vector of scaling coefficients}

\item{X}{matrix of size n by p contains the covariates}

\item{tol}{a small, non-negative number specifying the convergence
tolerance for the IBSS fitting procedure. The fitting procedure
will halt when the difference in the variational lower bound, or
\dQuote{ELBO} (the objective function to be maximized), is less
than \code{tol}.}

\item{low_wc}{vector of index of which wavelet coefficients display pathologic distribution}

\item{init_pi0_w}{starting value of weight on null compoenent in mixsqp
(between 0 and 1)}

\item{control_mixsqp}{list of parameter for mixsqp function see  mixsqp package}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{lowc_wc}{list of wavelet coefficients that exhibit too little variance}

\item{nullweight}{numeric value for penalizing likelihood at point mass 0 (should be between 0 and 1)
(usefull in small sample size)}

\item{cal_obj}{logical if set as TRUE compute ELBO for convergence monitoring}

\item{verbose}{If \code{verbose = TRUE}, the algorithm's progress,
and a summary of the optimization settings are printed to the
console.}

\item{cov_lev}{numeric between 0 and 1, corresponding to the
expected level of coverage of the cs if not specified set to 0.95}

\item{min.purity}{minimum purity for estimated credible sets}

\item{maxit}{Maximum number of IBSS iterations.}

\item{tt}{output of the cal_Bhat_Shat function}
}
\description{
Implementation of the SuSiF method
}
\details{
Implementation of the SuSiF method
}
