% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{test_stop_cond}
\alias{test_stop_cond}
\alias{test_stop_cond.susiF}
\title{Check tolerance for stopping criterion}
\usage{
test_stop_cond(susiF.obj, check, cal_obj, Y, X, D, C, indx_lst, ...)

\method{test_stop_cond}{susiF}(susiF.obj, check, cal_obj, Y, X, D, C, indx_lst, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by init_susiF_obj function}

\item{check}{numeric, dynamic value for testing outing of th while loop}

\item{cal_obj}{logical, if set to TRUE compute ELBO}

\item{X}{matrix of covariates}

\item{D}{matrix of wavelet D coefficients from the original input data (Y)}

\item{C}{vector of wavelet scaling coefficient from the original input data (Y)}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
a matrix of size N by size J of partial residuals
}
\description{
Check tolerance for stopping criterion
}
\keyword{internal}
