% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{TI_regression}
\alias{TI_regression}
\alias{TI_regression.susiF}
\title{Compute refined estimate using translation invariant wavelet transform}
\usage{
TI_regression(susiF.obj, Y, X, verbose, filter.number, family, ...)

\method{TI_regression}{susiF}(
  susiF.obj,
  Y,
  X,
  verbose = TRUE,
  filter.number = 10,
  family = "DaubLeAsymm",
  ...
)
}
\arguments{
\item{susiF.obj}{a susiF object}

\item{Y}{matrix of responses}

\item{X}{matrix containing the covariates}

\item{verbose}{logical}

\item{filter.number}{see wd description in wavethresh package description}

\item{family}{see wd description in wavethresh package description}

\item{\dots}{Other arguments.}
}
\description{
e Compute refined estimate using translation invariant wavelet transform
}
