% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_functions.R
\name{univariate_functional_regression}
\alias{univariate_functional_regression}
\title{Wrapper for univariate functional  regression used on susiF}
\usage{
univariate_functional_regression(
  Y,
  X,
  method = c("TI", "HMM", "smash"),
  filter.number = 1,
  verbose = FALSE,
  family = "DaubExPhase",
  alpha = 0.05,
  ...
)
}
\arguments{
\item{Y}{matrix of responses}

\item{X}{a one column matrix containing the covariate of interest}

\item{method}{method to be used, allowed values are "TI", "HMM","smash"}

\item{filter.number}{see wd description in wavethresh package description}

\item{verbose}{logical}

\item{family}{see wd description in wavethresh package description}

\item{alpha}{required confidence level}

\item{\dots}{Other arguments.}
}
\description{
Compute refined estimate using translation invariant wavelet transform
}
