% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_prior.R
\name{update_ash_pi}
\alias{update_ash_pi}
\title{Update ash object mixture proportion}
\usage{
update_ash_pi(G, tpi)
}
\arguments{
\item{tpi}{a vector of proportion}

\item{G_prior}{a ash object}
}
\value{
an ash object with updated mixture proportion
}
\description{
Update ash object mixture proportion
}
