% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_cal_credible_band}
\alias{update_cal_credible_band}
\alias{update_cal_credible_band.susiF}
\title{Update susiF by computing credible band for posterior curves}
\usage{
update_cal_credible_band(susiF.obj, indx_lst, ...)

\method{update_cal_credible_band}{susiF}(susiF.obj, indx_lst, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelet which assume that J is of the form J^2. If J not a power of 2, susiF internally remaps the data into grid of length 2^J}

\item{X}{matrix of size N by p}
}
\value{
susiF object
}
\description{
Update susiF by computing credible band for posterior curves
}
