% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_EBmvFR_obj.R,
%   R/operation_on_susiF_obj.R
\name{update_cal_fit_func.EBmvFR}
\alias{update_cal_fit_func.EBmvFR}
\alias{update_cal_fit_func}
\alias{update_cal_fit_func.susiF}
\title{Update susiF by computing posterior curves}
\usage{
\method{update_cal_fit_func}{EBmvFR}(EBmvFR.obj, indx_lst, ...)

update_cal_fit_func(susiF.obj, indx_lst, ...)

\method{update_cal_fit_func}{susiF}(susiF.obj, indx_lst, ...)
}
\arguments{
\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{susiF.obj}{a susiF object defined by init_susiF_obj function}
}
\value{
susiF object
}
\description{
Update susiF by computing posterior curves
}
\keyword{internal}
