% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_EBmvFR_obj.R,
%   R/operation_on_susiF_obj.R
\name{update_cal_fit_func.EBmvFR}
\alias{update_cal_fit_func.EBmvFR}
\alias{update_cal_fit_func}
\alias{update_cal_fit_func.susiF}
\title{Update susiF by computing posterior curves}
\usage{
\method{update_cal_fit_func}{EBmvFR}(EBmvFR.obj, indx_lst, ...)

update_cal_fit_func(
  susiF.obj,
  Y,
  X,
  indx_lst,
  TI,
  HMM,
  filter.number = 10,
  family = "DaubLeAsymm",
  ...
)

\method{update_cal_fit_func}{susiF}(
  susiF.obj,
  Y,
  X,
  indx_lst,
  TI = TRUE,
  HMM = FALSE,
  filter.number = 10,
  family = "DaubLeAsymm",
  ...
)
}
\arguments{
\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{susiF.obj}{a susiF object defined by init_susiF_obj function}

\item{Y}{functional phenotype, matrix of size N by size J. The underlying algorithm uses wavelet which assume that J is of the form J^2. If J not a power of 2, susif internally remaps the data into grid of length 2^J}

\item{X}{matrix of size n by p in}

\item{TI}{logical if set to true use cycle spinning wavelet regression}

\item{filter.number}{see documentation of wd from wavethresh package}

\item{family}{see documentation of wd from wavethresh package}
}
\value{
susiF object
}
\description{
Update susiF by computing posterior curves
}
\keyword{internal}
