% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_cal_fit_func}
\alias{update_cal_fit_func}
\alias{update_cal_fit_func.susiF}
\title{Update susiF by computing posterior curves}
\usage{
update_cal_fit_func(susiF.obj, indx_lst, ...)

\method{update_cal_fit_func}{susiF}(susiF.obj, indx_lst, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}
}
\value{
susiF object
}
\description{
Update susiF by computing posterior curves
}
\keyword{internal}
