% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{update_cal_pip}
\alias{update_cal_pip}
\alias{update_cal_pip.susiF}
\alias{update_cal_pip.susiF_ss}
\title{Update susiF by computing PiP}
\usage{
update_cal_pip(susiF.obj, ...)

\method{update_cal_pip}{susiF}(susiF.obj, ...)

\method{update_cal_pip}{susiF_ss}(susiF_ss.obj, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}
}
\value{
susiF object
}
\description{
Update susiF by computing PiP
}
