% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{update_ELBO}
\alias{update_ELBO}
\alias{update_ELBO.susiF}
\alias{update_ELBO.susiF_ss}
\title{Update susiF log Bayes factor}
\usage{
update_ELBO(susiF.obj, ELBO, ...)

\method{update_ELBO}{susiF}(susiF.obj, ELBO, ...)

\method{update_ELBO}{susiF_ss}(susiF_ss.obj, ELBO, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{ELBO}{new ELBO value}

\item{lBF}{vector of length p, containning the updated log Bayes factors}
}
\value{
susiF object
}
\description{
Update susiF log Bayes factor
}
