% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_lBF}
\alias{update_lBF}
\alias{update_lBF.susiF}
\title{Update susiF log Bayes factor}
\usage{
update_lBF(susiF.obj, l, lBF, ...)

\method{update_lBF}{susiF}(susiF.obj, l, lBF, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by init_susiF_obj function}

\item{l}{effect to update}

\item{lBF}{vector of length p, containing the updated log Bayes factors}
}
\value{
susiF object
}
\description{
Update susiF log Bayes factor
}
\keyword{internal}
