% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_lfsr}
\alias{update_lfsr}
\alias{update_lfsr.susiF}
\title{Update susiF local False Sign Rate}
\usage{
update_lfsr(susiF.obj, l, Bhat, Shat, indx_lst, ...)

\method{update_lfsr}{susiF}(susiF.obj, l, Bhat, Shat, indx_lst, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by init_susiF_obj function}

\item{l}{effect to update}

\item{Bhat}{matrix pxJ regression coefficient, Bhat[j,t] corresponds to regression coefficient of Y[,t] on X[,j]}

\item{Shat}{matrix pxJ standard error, Shat[j,t] corresponds to standard error of the regression coefficient of Y[,t] on X[,j]}

\item{indx_lst}{list generated by gen_wavelet_indx for the given level of resolution}

\item{alpha}{vector of length p, containing the updated log Bayes factors}
}
\value{
susiF object
}
\description{
Update susiF local False Sign Rate

Update susiF lfsr.
}
\keyword{internal}
