% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_lfsr_effect}
\alias{update_lfsr_effect}
\alias{update_lfsr_effect.susiF}
\title{Update susiF lfsr effect}
\usage{
update_lfsr_effect(susiF.obj, ...)

\method{update_lfsr_effect}{susiF}(susiF.obj)
}
\arguments{
\item{susiF.obj}{a susiF object defined by init_susiF_obj function}
}
\value{
susiF object
}
\description{
Update susiF lfsr effect
}
\keyword{internal}
