% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susiFutility.R
\name{update_prior.mixture_normal}
\alias{update_prior.mixture_normal}
\title{Update mixture proportion for mixture normal prior}
\usage{
\method{update_prior}{mixture_normal}(G_prior, tpi)

\method{update_prior}{mixture_normal}(G_prior, tpi)
}
\arguments{
\item{tpi}{a vector of proportion of class"pi_mixture_normal"}

\item{G}{a prior of class "mixture_normal"}
}
\value{
a prior of class "mixture_normal"

a prior of class "mixture_normal"
}
\description{
Update mixture proportion for mixture normal prior

Update mixture proportion for mixture normal prior
}
