% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{update_residual_variance}
\alias{update_residual_variance}
\title{Update residual variance}
\usage{
update_residual_variance(susiF.obj, sigma2)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{sigma2}{estimate residual variance}
}
\value{
updated susiF.obj
}
\description{
Update residual variance
}
