% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R,
%   R/operation_on_susiF_ss_obj.R
\name{update_susiF_obj}
\alias{update_susiF_obj}
\alias{update_susiF_obj.susiF}
\alias{update_susiF_obj.susiF_ss}
\title{Update  susiF object using the output of EM_pi}
\usage{
update_susiF_obj(susiF.obj, l, EM_pi, Bhat, Shat, indx_lst, ...)

\method{update_susiF_obj}{susiF}(
  susiF.obj,
  l,
  EM_pi,
  Bhat,
  Shat,
  indx_lst,
  cal_wc_lsfr = FALSE,
  ...
)

\method{update_susiF_obj}{susiF_ss}(susiF_ss.obj, l, EM_pi, Bhat, Shat, indx_lst, ...)
}
\arguments{
\item{susiF.obj}{a susiF object defined by \code{\link{init_susiF_obj}} function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}
}
\value{
susiF object
}
\description{
Update  susiF object using the output of EM_pi
}
