% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{update_susiF_obj}
\alias{update_susiF_obj}
\title{Update  susif object using output of EM}
\usage{
update_susiF_obj(susiF.obj, l, EM_pi, Bhat, Shat, indx_lst)
}
\arguments{
\item{susiF.obj}{a susisf object defined by init_susiF_bj function}

\item{l}{integer larger or equal to 1. Corresponds to the effect to be accessed}

\item{EM_pi}{an object of the class "EM_pi" generated by the function EM_pi}
}
\value{
susiF object
}
\description{
Update  susif object using output of EM
}
