% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{which_dummy_cs}
\alias{which_dummy_cs}
\alias{which_dummy_cs.susiF}
\title{Return which credible sets are  dummy}
\usage{
which_dummy_cs(obj, min.purity = 0.5, X, median_crit = FALSE, ...)

\method{which_dummy_cs}{susiF}(obj, min.purity = 0.5, X, median_crit = FALSE, ...)
}
\arguments{
\item{obj}{a susif object defined by init_susiF_obj function}

\item{min.purity}{minimal purity within a CS}

\item{X}{matrix of covariates}

\item{max_crit}{remove cs base on max absolute correlation instead of min absolute correlation, usefull in the}
}
\value{
a list of index corresponding the the dummy effect
}
\description{
Return which credible sets are  dummy
}
\keyword{internal}
