% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_on_susiF_obj.R
\name{which_dummy_cs}
\alias{which_dummy_cs}
\title{Return which credible sets are  dummy}
\usage{
which_dummy_cs(susiF.obj, min.purity = 0.5, X, ...)
}
\arguments{
\item{susiF.obj}{a susif object defined by \code{\link{init_susiF_obj}} function}

\item{min.purity}{minimal purity within a CS}

\item{X}{matrix of covariates}
}
\value{
a list of index corresponding the the dummy effect
}
\description{
Return which credible sets are  dummy
}
